refactor(#338): split WebEndpoints.cpp into focused web TUs
No behavior change — a pure mechanical move of the 2364-line monolith into
a TU family behind the unchanged WebEndpoints.h seam:

- WebEndpoints.cpp (456): shared state, buildCurrentSettingsJson, init
  (middleware + per-module registrars), the load-bearing-order loop drain,
  content snapshot + MQTT setters
- WebEndpointsInternal.h (77): internal seam — extern shared state,
  WebStateLock, registrar/loop hooks; include from Web*.cpp only
- WebContent.cpp (149): PROGMEM assets + SSE; sole WebAssets.h includer
  (/wifi-setup moved here — it serves a WebAssets blob)
- WebSettings.cpp (275): /settings, / POST, /reboot, wifi flow, onNotFound,
  /mqtt/discover + its mDNS drain
- WebSystem.cpp (304): /health /log /log/flash* /coredump/* /system/*
  /api /status
- WebFirmware.cpp (359): master OTA + rescue + rescue-boot + /debug/ota +
  the OTA stall watchdog as webFirmwareLoop()
- WebMaintenance.cpp (373): unit health + all {"seq":N} maintenance ops
- WebCluster.cpp (598): CORS/CSRF middleware + all /cluster/* routes +
  the cluster mDNS discovery drain

Route set verified identical (63 path+method pairs), drain order preserved
verbatim, 754 native + 71 pytest + S3 build green. CLAUDE.md references
updated.
