+++ firmware/v2/Master/WebContent.cpp
@@ -1,0 +1,149 @@
+// Static PROGMEM assets + the SSE display push (#251) — split from
+// WebEndpoints.cpp (#338); async-context rules in WebEndpoints.cpp's header.
+//
+// This TU is the ONLY WebAssets.h includer: the generated arrays have
+// internal linkage, so a second include would duplicate every PROGMEM blob.
+// Everything else reaches the bundled unit image through the
+// webUnitFirmwareBin()/webUnitFirmwareBinLen() accessors below, and the
+// portal page through the /wifi-setup route here.
+
+#include "WebEndpoints.h"
+#include "WebEndpointsInternal.h"
+
+#include <ESPAsyncWebServer.h>
+
+#include "ClusterLayout.h"  // CLUSTER_MAX_MEMBERS
+#include "ClusterLeader.h"
+#include "DisplayEvents.h"
+#include "Tasks.h"
+#include "WebAssets.h"
+
+// SSE display push (#251): netTask sends a "display" event the moment the
+// snapshot's text changes (webDisplayEventsTick below); onConnect hands a
+// fresh client the current state so it paints without waiting for a
+// change. All send-side calls come from netTask; the esp32async event
+// source queues per-client internally.
+static AsyncEventSource sseEvents("/events");
+
+// Serves a gzipped PROGMEM asset (v1 #159). Cache policy: PROGMEM assets
+// only change via a master OTA, so the browser must revalidate every
+// navigation or an OTA that swaps the UI would leave stale HTML/JS behind.
+static void serveGzipAsset(AsyncWebServerRequest* request,
+                           const char* contentType, const uint8_t* asset,
+                           size_t assetLen) {
+  AsyncWebServerResponse* resp =
+      request->beginResponse(200, contentType, asset, assetLen);
+  resp->addHeader("Content-Encoding", "gzip");
+  resp->addHeader("Cache-Control", "no-cache");
+  request->send(resp);
+}
+
+// Wall-aware /events payload (#277): while leading a cluster, the display
+// event carries every reconstructed grid row; rowsKeyOut feeds the tick's
+// change tracker. Locks run strictly sequentially — the content snapshot
+// (webStateMutex) is taken and RELEASED before the leader mutex, never
+// nested.
+static String sseDisplayPayload(const DisplaySnapshot& snap,
+                                String* rowsKeyOut) {
+  String rows[CLUSTER_MAX_MEMBERS];
+  int selfRow = 0;
+  int rowCount = 0;
+  if (clusterLeaderEnabled()) {
+    WebContentSnapshot content = webDisplayContentSnapshot();
+    rowCount = clusterLeaderMirrorRows(rows, selfRow, String(snap.currentText),
+                                       content.alignment);
+  }
+  if (rowsKeyOut != nullptr) *rowsKeyOut = displayEventRowsKey(rows, rowCount);
+  return buildDisplayEventJson(snap.currentText, rows, rowCount, selfRow);
+}
+
+void webContentRegister(AsyncWebServer& server) {
+  // --- SSE (#251) ----------------------------------------------------------
+  // A fresh client gets the current display text immediately; every later
+  // change is pushed by webDisplayEventsTick() from netTask.
+  sseEvents.onConnect([](AsyncEventSourceClient* client) {
+    DisplaySnapshot snap = displaySnapshotGet();
+    client->send(sseDisplayPayload(snap, nullptr).c_str(), "display",
+                 millis());
+  });
+  server.addHandler(&sseEvents);
+
+  // --- static assets, all from PROGMEM (generated by build_assets.py) -----
+  server.on("/", HTTP_GET, [](AsyncWebServerRequest* request) {
+    serveGzipAsset(request, "text/html", INDEX_HTML_GZ, INDEX_HTML_GZ_LEN);
+  });
+  server.on("/index.html", HTTP_GET, [](AsyncWebServerRequest* request) {
+    serveGzipAsset(request, "text/html", INDEX_HTML_GZ, INDEX_HTML_GZ_LEN);
+  });
+  server.on("/script.js", HTTP_GET, [](AsyncWebServerRequest* request) {
+    serveGzipAsset(request, "application/javascript", SCRIPT_JS_GZ,
+                   SCRIPT_JS_GZ_LEN);
+  });
+  server.on("/md5.js", HTTP_GET, [](AsyncWebServerRequest* request) {
+    serveGzipAsset(request, "application/javascript", MD5_JS_GZ, MD5_JS_GZ_LEN);
+  });
+  server.on("/style.css", HTTP_GET, [](AsyncWebServerRequest* request) {
+    serveGzipAsset(request, "text/css", STYLE_CSS_GZ, STYLE_CSS_GZ_LEN);
+  });
+  server.on("/favicon.png", HTTP_GET, [](AsyncWebServerRequest* request) {
+    request->send(request->beginResponse(200, "image/png", FAVICON_PNG,
+                                         FAVICON_PNG_LEN));
+  });
+  // Full IANA→POSIX timezone table (#252), baked from the vendored
+  // posix_tz_db zones.csv; the Settings type-ahead fetches it lazily.
+  server.on("/tz.json", HTTP_GET, [](AsyncWebServerRequest* request) {
+    serveGzipAsset(request, "application/json", TZ_JSON_GZ, TZ_JSON_GZ_LEN);
+  });
+
+  // Portal page asset (#188) — served unconditionally; from the LAN it
+  // doubles as a "move to another network" page. The wifi flow handlers
+  // (/wifi/scan, /wifi/config, /reset-wifi) live in WebSettings.cpp; the
+  // route is here because the page is a WebAssets.h blob.
+  server.on("/wifi-setup", HTTP_GET, [](AsyncWebServerRequest* request) {
+    serveGzipAsset(request, "text/html", PORTAL_HTML_GZ, PORTAL_HTML_GZ_LEN);
+  });
+}
+
+void webDisplayEventsTick() {
+  // 100 ms check cadence: imperceptible latency, and the snapshot copy
+  // stays off netTask's 10 ms hot loop.
+  static uint32_t nextCheckMs = 0;
+  static DisplayEventTracker tracker;
+  uint32_t nowMs = millis();
+  if ((int32_t)(nowMs - nextCheckMs) < 0) return;
+  nextCheckMs = nowMs + 100;
+
+  // count() before consuming the change: the library runs onConnect before
+  // inserting the client into its list, so a change landing in that window
+  // must stay unconsumed for the next tick. The resulting duplicate push of
+  // an already-shown text is deduped by the mirror's frame compare.
+  if (sseEvents.count() == 0) return;
+  DisplaySnapshot snap = displaySnapshotGet();
+
+  // Cheap pre-check before paying for the wall reconstruction (mutex +
+  // String churn, every 100 ms otherwise): only rebuild when the own text
+  // changed, the grid generation moved, or leading flipped. The rows-key
+  // due-check below stays authoritative for what actually gets pushed.
+  static uint32_t lastGridGen = 0;
+  static bool lastLeading = false;
+  bool leading = clusterLeaderEnabled();
+  uint32_t gridGen = leading ? clusterLeaderGridGeneration() : 0;
+  if (strncmp(tracker.lastText, snap.currentText, DISPLAY_CMD_TEXT_LEN) == 0 &&
+      leading == lastLeading && gridGen == lastGridGen) {
+    return;
+  }
+  lastLeading = leading;
+  lastGridGen = gridGen;
+
+  String rowsKey;
+  String payload = sseDisplayPayload(snap, &rowsKey);
+  if (!displayEventDue(tracker, snap.currentText, rowsKey)) return;
+  sseEvents.send(payload.c_str(), "display", millis());
+}
+
+// Bundled unit firmware accessors (#205) — see the header comment above for
+// why this TU is the only WebAssets.h includer. On the S3, PROGMEM is
+// flash-mapped and directly readable, so the pointer works as a plain byte
+// buffer.
+const uint8_t* webUnitFirmwareBin() { return UNIT_FIRMWARE_BIN; }
+size_t webUnitFirmwareBinLen() { return UNIT_FIRMWARE_BIN_LEN; }
