+++ firmware/v2/Master/WebEndpoints.cpp
@@ -1,5 +1,10 @@
-// Web endpoint registration for the v2 master (#186) — port of v1's
-// ServiceWebEndpoints.ino onto the ESP32Async server stack.
+// Web endpoint core for the v2 master (#186) — shared state, init/loop
+// lifecycle and the cross-task accessors. The route handlers live in the
+// Web*.cpp family (#338 split): WebContent (assets/SSE), WebSettings
+// (settings/wifi/mqtt-discover), WebSystem (diagnostics reads), WebFirmware
+// (master OTA/rescue), WebMaintenance (unit ops), WebCluster (cluster wire +
+// CORS middleware). Shared internals cross that family through
+// WebEndpointsInternal.h ONLY.
 //
 // Async-context rule (v1 #150) carries over verbatim: handlers run in the
 // async_tcp task. They may parse, validate, read state and respond — they
@@ -8,132 +13,47 @@
 // loop() via webEndpointsLoop().
 
 #include "WebEndpoints.h"
+#include "WebEndpointsInternal.h"
 
 #include <ESPAsyncWebServer.h>
-#include <LittleFS.h>
-#include <WiFi.h>  // localIP in the leading-board join refusal (#295)
-#include <Update.h>
-#include <freertos/FreeRTOS.h>
-#include <freertos/semphr.h>
+#include <WiFi.h>
 
-#include <memory>
-
-#include <esp_partition.h>
-#include <esp_core_dump.h>  // #319: /coredump crash diagnostics
-
-#include "ApiIndex.h"
 #include "BuildVersion.h"
 #include "ClockPolicy.h"
 #include "ClockService.h"
-#include "ClusterDigest.h"
-#include "ClusterHmac.h"  // #313 follow-on: rebuild canonical msgs for verify
-#include "ClusterDiscovery.h"
+#include "ClusterDigest.h"  // clusterCsrfRejectPost (inline upload gate)
 #include "ClusterFollower.h"
-#include "ClusterLayout.h"  // CLUSTER_MAX_MEMBERS / CLUSTER_HOST_MAX_LEN
 #include "ClusterLeader.h"
-#include "DisplayEvents.h"
-#include "FactorySlot.h"
-#include "FollowerImagePolicy.h"  // #304 follower-image upload guard
-#include "FollowerImageStore.h"
 #include "FlashLog.h"
+#include "FollowerImageStore.h"
 #include "HelpersSerialHandling.h"
-#include "MaintenancePolicy.h"
-#include "OtaService.h"
-#include "PendingSettingsPost.h"
-#include "SettingsJson.h"
-#include "MdnsDiscovery.h"
 #include "MqttService.h"
+#include "OtaService.h"
+#include "SettingsJson.h"
 #include "SplitFlapProtocol.h"
-#include "SystemStats.h"
-#include "SystemStatsPolicy.h"
 #include "Tasks.h"
-#include "UnitBus.h"
-#include "WearPolicy.h"
-#include "WebAssets.h"
-#include "WebLog.h"
-#include "WifiService.h"
 
-#include <ESPmDNS.h>
-
-// SSE display push (#251): netTask sends a "display" event the moment the
-// snapshot's text changes (webDisplayEventsTick below); onConnect hands a
-// fresh client the current state so it paints without waiting for a
-// change. All send-side calls come from netTask; the esp32async event
-// source queues per-client internally.
-static AsyncEventSource sseEvents("/events");
-
-// Staged mutations, owned here; drained by webEndpointsLoop().
-static PendingSettingsPost pendingPost;
-static bool pendingReboot = false;
-static uint32_t rebootRequestedAtMs = 0;
-static String pendingIntendedVersion;  // ?v= from /firmware/master (#190)
-static bool pendingIntendedVersionProvided = false;
-
-// OTA upload rejection state (v1 ServiceBootModes pattern): the upload
-// callback can't respond, so it records the rejection and the completion
-// callback sends it. Both callbacks run in the async_tcp task — same task,
-// no cross-task race, no mutex (uploads are serialized by the TCP stream).
-static int otaRejectionStatus = 0;  // 0 = not rejected
-static String otaRejectionReason;
-
-// Concurrent-upload guard (#191): the request that owns the live Update
-// session. A second overlapping POST must not abort/hijack it — the
-// overlapping request is marked rejected via its _tempObject (freed by the
-// request destructor) and answered 409, and it never touches the shared
-// rejection state above, which belongs to the owner. Same async_tcp task
-// for all callbacks — no lock needed. Compared only, never dereferenced.
-static AsyncWebServerRequest* otaOwnerRequest = nullptr;
-
-// Upload throughput measurement (#248): decides whether flash erase or the
-// network dominates OTA wall time before any speed work is designed. Owned
-// by the async_tcp task like the session state above.
-static uint32_t otaUploadStartMs = 0;
-
-// Stall watchdog (#313): last time a chunk of the live OTA arrived, stamped
-// per-chunk. A slow-loris that opens /firmware/master and then stops (no
-// preflight, MQTT frozen for the flash) would otherwise wedge OTA + MQTT
-// indefinitely. webEndpointsLoop() (netTask) aborts a session idle past the
-// timeout — the port of the ESP-01 follower's webOtaUploadFrozen(). Written
-// by the async_tcp task, read (and cleared via the session) by netTask; a
-// stalled session has no concurrent writer, so the plain uint32 is safe.
-static uint32_t otaLastChunkMs = 0;
-static const uint32_t OTA_STALL_TIMEOUT_MS = 30000UL;
-
-// Same pattern for POST /firmware/rescue (#195). Separate state on purpose:
-// a rescue install and a master OTA are different flows and must not read
-// each other's leftovers. (Concurrent uploads remain #191 territory.)
-static int rescueRejectionStatus = 0;
-static String rescueRejectionReason;
-
-// Same pattern again for POST /cluster/follower-firmware (#304) — the stored
-// ESP-01 image upload. Independent of the OTA/rescue flows above.
-static int followerFwRejectStatus = 0;
-static String followerFwRejectReason;
+// Staged mutations, owned here; drained by webEndpointsLoop(). External
+// linkage across the Web*.cpp family (WebEndpointsInternal.h).
+PendingSettingsPost pendingPost;
+bool pendingReboot = false;
+uint32_t rebootRequestedAtMs = 0;
+String pendingIntendedVersion;  // ?v= from /firmware/master (#190)
+bool pendingIntendedVersionProvided = false;
 
 // Live state the read handlers render. Set once in webEndpointsInit();
-// handlers only ever read (async-context rule). Held as TU-local statics
+// handlers only ever read (async-context rule). Held as module globals
 // rather than lambda-captured references so a handler can never outlive
 // what it captured.
-static MasterSettings* liveSettings = nullptr;
-static SettingsStore* liveStore = nullptr;  // MQTT setters persist through it
-static String effectiveName;
-
-// MQTT broker mDNS discovery staging (#224, v1 /mqtt/discover contract):
-// the POST arms the flag, netTask's drain runs the blocking query and
-// caches the JSON, the GET answers 202 while pending / 200 from the cache.
-static bool mqttDiscoverPending = false;
-static String mqttDiscoverResultJson;
-
-// Cluster board discovery staging (#274): same POST-arm / netTask-drain /
-// GET-poll contract as the MQTT pair above, browsing _splitflap._tcp.
-static bool clusterDiscoverPending = false;
-static String clusterDiscoverResultJson;
+MasterSettings* liveSettings = nullptr;
+SettingsStore* liveStore = nullptr;  // MQTT setters persist through it
+String effectiveName;
 
 // Runtime-only message state (#192, v1 parity: never persisted, "" at
 // boot). Written by the drain, read by GET /settings and the clock ticker's
 // webDisplayContentSnapshot() — all under webStateMutex.
-static String currentInputText;
-static String lastMessageStamp;
+String currentInputText;
+String lastMessageStamp;
 
 // Cross-task guard. Unlike v1's single-core cooperative ESP8266, the
 // handlers here run in the async_tcp FreeRTOS task while the drain runs in
@@ -142,100 +62,7 @@ static String lastMessageStamp;
 // mid-free. A mutex (not a spinlock) on purpose: the critical sections
 // allocate Strings and the drain writes NVS, neither of which is allowed
 // inside portENTER_CRITICAL.
-static SemaphoreHandle_t webStateMutex = nullptr;
-
-struct WebStateLock {
-  WebStateLock() { xSemaphoreTake(webStateMutex, portMAX_DELAY); }
-  ~WebStateLock() { xSemaphoreGive(webStateMutex); }
-  WebStateLock(const WebStateLock&) = delete;
-  WebStateLock& operator=(const WebStateLock&) = delete;
-};
-
-// Serves a gzipped PROGMEM asset (v1 #159). Cache policy: PROGMEM assets
-// only change via a master OTA, so the browser must revalidate every
-// navigation or an OTA that swaps the UI would leave stale HTML/JS behind.
-static void serveGzipAsset(AsyncWebServerRequest* request,
-                           const char* contentType, const uint8_t* asset,
-                           size_t assetLen) {
-  AsyncWebServerResponse* resp =
-      request->beginResponse(200, contentType, asset, assetLen);
-  resp->addHeader("Content-Encoding", "gzip");
-  resp->addHeader("Cache-Control", "no-cache");
-  request->send(resp);
-}
-
-// --- maintenance endpoint helpers (#204) ---------------------------------------
-
-// v1 parseCalibrationAddress as a seam: policy verdicts come from the pure
-// MaintenancePolicy.h, this only translates request → verdict → response.
-// Validates against the caller's snapshot COPY — a fast, possibly stale
-// view; displayTask re-runs whatever check the queue delay can invalidate.
-static bool maintCheckAddress(AsyncWebServerRequest* request,
-                              const DisplaySnapshot& snap, int& outAddr) {
-  String raw;
-  bool provided = request->hasParam("address");
-  if (provided) raw = request->getParam("address")->value();
-  MaintVerdict verdict = maintValidateAddress(provided ? raw.c_str() : nullptr,
-                                              snap.units, UNITS_AMOUNT,
-                                              outAddr);
-  if (verdict.httpStatus != 200) {
-    request->send(verdict.httpStatus, "text/plain", verdict.message);
-    return false;
-  }
-  return true;
-}
-
-// Required numeric query param (strtol base 0, v1 hex support). Sends the
-// 400 itself so callers just early-return.
-static bool maintRequireLongParam(AsyncWebServerRequest* request,
-                                  const char* name, long& out) {
-  if (!request->hasParam(name)) {
-    String msg = "Missing '";
-    msg += name;
-    msg += "' query param";
-    request->send(400, "text/plain", msg);
-    return false;
-  }
-  String raw = request->getParam(name)->value();
-  char* end = nullptr;
-  out = strtol(raw.c_str(), &end, 0);
-  if (end == raw.c_str()) {
-    String msg = "'";
-    msg += name;
-    msg += "' must be a number";
-    request->send(400, "text/plain", msg);
-    return false;
-  }
-  return true;
-}
-
-// Producer gate (#205): while a reflash job runs, display-mutating POSTs
-// answer 409 instead of enqueueing — nothing piles up behind the job to
-// burst-drain afterwards. /stop is the ONE exception (it is the cancel)
-// and deliberately does not call this.
-static bool rejectWhileReflashing(AsyncWebServerRequest* request) {
-  if (!reflashInProgress(displaySnapshotGet().reflash)) return false;
-  request->send(409, "text/plain",
-                F("Unit reflash in progress — retry when it finishes"));
-  return true;
-}
-
-// Enqueue-or-503 tail shared by every maintenance POST: the client either
-// gets its correlation seq or an honest busy — never a silently dropped op.
-// Every caller except /stop funnels through here, so the reflash gate
-// lives here too.
-static void maintEnqueue(AsyncWebServerRequest* request,
-                         const DisplayCommand& cmd) {
-  if (rejectWhileReflashing(request)) return;
-  if (!displayEnqueue(cmd)) {
-    request->send(503, "text/plain",
-                  F("Display queue full — try again in a moment"));
-    return;
-  }
-  char buf[24];
-  snprintf(buf, sizeof(buf), "{\"seq\":%lu}", (unsigned long)cmd.seq);
-  request->send(200, "application/json", buf);
-}
+SemaphoreHandle_t webStateMutex = nullptr;
 
 const char* webResetReasonString() {
   switch (esp_reset_reason()) {
@@ -253,63 +80,13 @@ const char* webResetReasonString() {
   }
 }
 
-// Wall-aware /events payload (#277): while leading a cluster, the display
-// event carries every reconstructed grid row; rowsKeyOut feeds the tick's
-// change tracker. Locks run strictly sequentially — the content snapshot
-// (webStateMutex) is taken and RELEASED before the leader mutex, never
-// nested.
-static String sseDisplayPayload(const DisplaySnapshot& snap,
-                                String* rowsKeyOut) {
-  String rows[CLUSTER_MAX_MEMBERS];
-  int selfRow = 0;
-  int rowCount = 0;
-  if (clusterLeaderEnabled()) {
-    WebContentSnapshot content = webDisplayContentSnapshot();
-    rowCount = clusterLeaderMirrorRows(rows, selfRow, String(snap.currentText),
-                                       content.alignment);
-  }
-  if (rowsKeyOut != nullptr) *rowsKeyOut = displayEventRowsKey(rows, rowCount);
-  return buildDisplayEventJson(snap.currentText, rows, rowCount, selfRow);
-}
-
-// #294 rung 3 + #313 CSRF gate. Two jobs, one attach point (both gates are
-// pure ClusterDigest.h logic; simple requests only — form posts / GETs, no
-// preflight handler needed):
-//   1. ENFORCE (#313): a mutating POST carrying a browser Origin that is not
-//      a LAN pane is cross-site forgery — 403 BEFORE the handler runs, so
-//      the whole class of mutating routes (/firmware/master, /reset-wifi,
-//      /cluster/*, …) is covered without a per-route allowlist to drift.
-//      The leader's esp_http_client sends no Origin and passes; the board's
-//      own LAN UI sends a LAN origin and passes.
-//   2. REFLECT (#294): on the per-member management surface, echo the LAN
-//      origin back so another pane's browser can read the response.
-static AsyncMiddlewareFunction clusterCorsMiddleware(
-    [](AsyncWebServerRequest* request, ArMiddlewareNext next) {
-      bool hasOrigin = request->hasHeader("Origin");
-      String origin = hasOrigin ? request->header("Origin") : String();
-      if (clusterCsrfRejectPost(request->method() == HTTP_POST, hasOrigin,
-                                origin)) {
-        request->send(403, "text/plain",
-                      F("Cross-origin POST refused (CSRF guard)"));
-        return;  // handler chain stops — next() is never called
-      }
-      next();
-      if (!hasOrigin) return;
-      if (!clusterCorsPathAllowed(request->url())) return;
-      if (!clusterCorsOriginAllowed(origin)) return;
-      AsyncWebServerResponse* response = request->getResponse();
-      if (response == nullptr) return;
-      response->addHeader("Access-Control-Allow-Origin", origin);
-      response->addHeader("Vary", "Origin");
-    });
-
-// #313: the server middleware above fires only at PARSE_REQ_END — AFTER an
+// #313: the CORS/CSRF server middleware fires only at PARSE_REQ_END — AFTER an
 // upload route's onUpload callback has already streamed the body to flash. So
 // the routes that WRITE inside onUpload (master OTA, rescue install, follower
 // image) must gate CSRF INLINE at index==0, before the first write, exactly
 // as the ESP-01 follower does. Returns true when the upload is a forged
 // cross-site POST; the caller marks its own per-request rejection state.
-static bool webUploadCsrfRejected(AsyncWebServerRequest* request) {
+bool webUploadCsrfRejected(AsyncWebServerRequest* request) {
   bool hasOrigin = request->hasHeader("Origin");
   return clusterCsrfRejectPost(true, hasOrigin,
                                hasOrigin ? request->header("Origin") : String());
@@ -318,7 +95,7 @@ static bool webUploadCsrfRejected(AsyncWebServerRequest* request) {
 // Gathers the full /settings JSON from the display snapshot, live settings and
 // the OTA/cluster/MQTT services. Extracted so both GET /settings and the
 // /status one-shot aggregate (#307) render the identical object.
-static String buildCurrentSettingsJson() {
+String buildCurrentSettingsJson() {
   SettingsJsonFields f;
   DisplaySnapshot snap = displaySnapshotGet();
   int addrs[UNITS_AMOUNT];
@@ -386,7 +163,7 @@ static String buildCurrentSettingsJson() {
 void webEndpointsInit(AsyncWebServer& server, MasterSettings& settings,
                       SettingsStore& store,
                       const String& effectiveDeviceName) {
-  server.addMiddleware(&clusterCorsMiddleware);
+  server.addMiddleware(&webClusterCorsMiddleware());
   // Handlers never write the store; the loop drain and the mqttTask-called
   // setters below do (both hold webStateMutex).
   webStateMutex = xSemaphoreCreateMutex();
@@ -400,1576 +177,15 @@ void webEndpointsInit(AsyncWebServer& server, MasterSettings& settings,
   liveStore = &store;
   effectiveName = effectiveDeviceName;
 
-  // --- SSE (#251) ----------------------------------------------------------
-  // A fresh client gets the current display text immediately; every later
-  // change is pushed by webDisplayEventsTick() from netTask.
-  sseEvents.onConnect([](AsyncEventSourceClient* client) {
-    DisplaySnapshot snap = displaySnapshotGet();
-    client->send(sseDisplayPayload(snap, nullptr).c_str(), "display",
-                 millis());
-  });
-  server.addHandler(&sseEvents);
-
-  // --- static assets, all from PROGMEM (generated by build_assets.py) -----
-  server.on("/", HTTP_GET, [](AsyncWebServerRequest* request) {
-    serveGzipAsset(request, "text/html", INDEX_HTML_GZ, INDEX_HTML_GZ_LEN);
-  });
-  server.on("/index.html", HTTP_GET, [](AsyncWebServerRequest* request) {
-    serveGzipAsset(request, "text/html", INDEX_HTML_GZ, INDEX_HTML_GZ_LEN);
-  });
-  server.on("/script.js", HTTP_GET, [](AsyncWebServerRequest* request) {
-    serveGzipAsset(request, "application/javascript", SCRIPT_JS_GZ,
-                   SCRIPT_JS_GZ_LEN);
-  });
-  server.on("/md5.js", HTTP_GET, [](AsyncWebServerRequest* request) {
-    serveGzipAsset(request, "application/javascript", MD5_JS_GZ, MD5_JS_GZ_LEN);
-  });
-  server.on("/style.css", HTTP_GET, [](AsyncWebServerRequest* request) {
-    serveGzipAsset(request, "text/css", STYLE_CSS_GZ, STYLE_CSS_GZ_LEN);
-  });
-  server.on("/favicon.png", HTTP_GET, [](AsyncWebServerRequest* request) {
-    request->send(request->beginResponse(200, "image/png", FAVICON_PNG,
-                                         FAVICON_PNG_LEN));
-  });
-  // Full IANA→POSIX timezone table (#252), baked from the vendored
-  // posix_tz_db zones.csv; the Settings type-ahead fetches it lazily.
-  server.on("/tz.json", HTTP_GET, [](AsyncWebServerRequest* request) {
-    serveGzipAsset(request, "application/json", TZ_JSON_GZ, TZ_JSON_GZ_LEN);
-  });
-
-  // --- reads ---------------------------------------------------------------
-  server.on("/settings", HTTP_GET, [](AsyncWebServerRequest* request) {
-    // Full gather extracted to buildCurrentSettingsJson() so GET /status
-    // renders the identical settings object (#307).
-    request->send(200, "application/json", buildCurrentSettingsJson());
-  });
-
-  server.on("/health", HTTP_GET, [](AsyncWebServerRequest* request) {
-    request->send(200, "text/plain", "Healthy");
-  });
-
-  server.on("/log", HTTP_GET, [](AsyncWebServerRequest* request) {
-    // Don't SerialPrintln here; every log request would otherwise stamp
-    // itself into the buffer on every poll and drown out real activity.
-    request->send(200, "text/plain", webLogRead());
-  });
-
-  // --- persistent flash log (#206) ------------------------------------------
-  // Serves the LittleFS log files directly (chunked by the async layer;
-  // esp_littlefs serializes fs access internally, and the flush path
-  // open→append→closes so no write handle is ever shared). Known-benign
-  // race: between the exists() check and the response's real open(), a
-  // rotation or drained clear can remove the file — AsyncFileResponse then
-  // degrades to a 404, so a request racing a rotate occasionally 404s;
-  // retry. ?prev=1 = the rotated file. Content up to the last ~5 s flush.
-  server.on("/log/flash", HTTP_GET, [](AsyncWebServerRequest* request) {
-    if (!flashLogAvailable()) {
-      request->send(503, "text/plain",
-                    F("Flash log unavailable (storage mount failed)"));
-      return;
-    }
-    const char* path = request->hasParam("prev") ? flashLogPreviousPath()
-                                                 : flashLogCurrentPath();
-    if (!LittleFS.exists(path)) {
-      request->send(404, "text/plain", F("No flash log yet"));
-      return;
-    }
-    request->send(LittleFS, path, "text/plain");
-  });
-
-  // Clear is staged and drained by netTask's flashLogTick — handlers never
-  // write flash (async rule).
-  server.on("/log/flash/clear", HTTP_POST, [](AsyncWebServerRequest* request) {
-    if (!flashLogAvailable()) {
-      request->send(503, "text/plain",
-                    F("Flash log unavailable (storage mount failed)"));
-      return;
-    }
-    flashLogRequestClear();
-    request->send(202, "text/plain", F("Flash log clear queued"));
-  });
-
-  // --- coredump (#319) — remote crash diagnostics --------------------------
-  // The SUMMARY (task name + code addresses + backtrace PCs) carries NO
-  // secrets, so an unauthenticated GET is fine — same posture as /settings,
-  // /log/flash. It is always registered. The RAW image, in contrast, is a
-  // task-stack dump that CAN transiently hold HMAC key material / WiFi-cred
-  // fragments (a key mid-sign on clusterTask's stack); this device has no
-  // auth and CORS-closing does not stop a plain HTTP client, so raw is
-  // compiled out of shipped images and only appears in a build that opts in
-  // with -DCOREDUMP_RAW_ENABLE (deliberately OTA'd to a crashing board).
-  // Read-only; the coredump partition is written solely by the panic handler.
-  server.on("/coredump/summary", HTTP_GET, [](AsyncWebServerRequest* request) {
-    if (esp_core_dump_image_check() != ESP_OK) {
-      request->send(404, "application/json", F("{\"present\":false}"));
-      return;
-    }
-    // Zero-init: if a partially-corrupted-but-parseable dump leaves a field
-    // untouched, we echo a deterministic 0, not stale heap bytes.
-    auto* summary =
-        (esp_core_dump_summary_t*)calloc(1, sizeof(esp_core_dump_summary_t));
-    if (summary == nullptr) {
-      request->send(503, "application/json", F("{\"error\":\"oom\"}"));
-      return;
-    }
-    if (esp_core_dump_get_summary(summary) != ESP_OK) {
-      free(summary);
-      request->send(500, "application/json",
-                    F("{\"present\":true,\"error\":\"summary-failed\"}"));
-      return;
-    }
-    char task[17];
-    memcpy(task, summary->exc_task, 16);
-    task[16] = '\0';
-    char hex[11];
-    String out;
-    out.reserve(640);
-    out += F("{\"present\":true,\"rev\":\"" GIT_REV "\",\"task\":");
-    appendJsonString(out, String(task));
-    snprintf(hex, sizeof(hex), "0x%08x", (unsigned)summary->exc_pc);
-    out += F(",\"pc\":\"");
-    out += hex;
-    out += F("\",\"excCause\":");
-    out += (unsigned)summary->ex_info.exc_cause;
-    snprintf(hex, sizeof(hex), "0x%08x", (unsigned)summary->ex_info.exc_vaddr);
-    out += F(",\"excVaddr\":\"");
-    out += hex;
-    out += F("\",\"coreDumpVersion\":");
-    out += (unsigned)summary->core_dump_version;
-    out += F(",\"backtraceCorrupted\":");
-    out += summary->exc_bt_info.corrupted ? F("true") : F("false");
-    out += F(",\"backtrace\":[");
-    uint32_t depth = summary->exc_bt_info.depth;
-    if (depth > 16) depth = 16;
-    for (uint32_t i = 0; i < depth; i++) {
-      if (i) out += ',';
-      snprintf(hex, sizeof(hex), "0x%08x", (unsigned)summary->exc_bt_info.bt[i]);
-      out += '"';
-      out += hex;
-      out += '"';
-    }
-    out += F("]}");
-    free(summary);
-    request->send(200, "application/json", out);
-  });
-
-#ifdef COREDUMP_RAW_ENABLE
-  // Raw ELF coredump for offline `esp-coredump info_corefile` — the fallback
-  // when a summary backtrace reads corrupted. Streamed straight off the
-  // coredump partition (flash READ only, async-safe). Opt-in only (see the
-  // summary comment above): task stacks can leak transient secrets, so this
-  // never ships in a release image.
-  server.on("/coredump/raw", HTTP_GET, [](AsyncWebServerRequest* request) {
-    if (esp_core_dump_image_check() != ESP_OK) {
-      request->send(404, "text/plain", F("no coredump"));
-      return;
-    }
-    size_t addr = 0, size = 0;
-    if (esp_core_dump_image_get(&addr, &size) != ESP_OK || size == 0) {
-      request->send(500, "text/plain", F("coredump image unavailable"));
-      return;
-    }
-    const esp_partition_t* part = esp_partition_find_first(
-        ESP_PARTITION_TYPE_DATA, ESP_PARTITION_SUBTYPE_DATA_COREDUMP, nullptr);
-    if (part == nullptr) {
-      request->send(500, "text/plain", F("no coredump partition"));
-      return;
-    }
-    // image_get() returns an ABSOLUTE flash address — translate to a
-    // partition-relative offset for esp_partition_read.
-    size_t offset = addr - part->address;
-    AsyncWebServerResponse* response = request->beginChunkedResponse(
-        "application/octet-stream",
-        [part, offset, size](uint8_t* buffer, size_t maxLen,
-                             size_t index) -> size_t {
-          if (index >= size) return 0;
-          size_t remaining = size - index;
-          size_t toRead = remaining < maxLen ? remaining : maxLen;
-          if (esp_partition_read(part, offset + index, buffer, toRead) !=
-              ESP_OK) {
-            return 0;  // abort the stream
-          }
-          return toRead;
-        });
-    response->addHeader("Content-Disposition",
-                        "attachment; filename=\"coredump-" GIT_REV ".elf\"");
-    request->send(response);
-  });
-#endif  // COREDUMP_RAW_ENABLE
-
-  // --- settings/message form (v1 wire contract) ----------------------------
-  server.on("/", HTTP_POST, [](AsyncWebServerRequest* request) {
-    // All-or-nothing per POST: parse into a local post, merge only when
-    // every provided field validated (v1 #128/#153 semantics, pure logic
-    // in PendingSettingsPost.h — natively tested).
-    PendingSettingsPost local;
-    bool submissionError = false;
-
-    int params = request->params();
-    for (int i = 0; i < params; i++) {
-      const AsyncWebParameter* p = request->getParam(i);
-      if (!p->isPost()) continue;
-      SettingsParamResult result =
-          stageSettingsParam(local, p->name(), p->value());
-      if (result == SettingsParamResult::Invalid) {
-        SerialPrintln("Invalid value for '" + p->name() + "': " + p->value());
-        submissionError = true;
-      }
-    }
-    if (!settingsPostConsistent(local)) {
-      SerialPrintln(F("Transient dwell provided without transient text."));
-      submissionError = true;
-    }
-
-    // Per-card fetch() saves (#128) send ajax=1 and want a status code
-    // instead of the classic redirect.
-    bool isAjax = request->hasParam("ajax", true);
-
-    if (submissionError) {
-      if (isAjax) request->send(400, "text/plain", F("invalid"));
-      else request->redirect("/?invalid-submission=true");
-      return;
-    }
-
-    // Message/transient sends become display commands at drain time; report
-    // a full queue now instead of accepting one that would be dropped. (The
-    // stub worker drains instantly, so this only fires if something wedges.)
-    // The reflash gate (#205) applies only to the display-bound part — pure
-    // settings saves don't touch the display queue and stay allowed.
-    if ((local.inputTextProvided || local.transientTextProvided) &&
-        reflashInProgress(displaySnapshotGet().reflash)) {
-      if (isAjax) {
-        request->send(409, "text/plain", F("reflash in progress"));
-      } else {
-        request->redirect("/?display-busy=true");
-      }
-      return;
-    }
-    // Cluster producer gate (#272): a clustered follower's text/mode belong
-    // to the leader — 409; the banner explains why. Transients stay allowed
-    // (they are the calibration vehicle — maintenance is local), and so do
-    // pure settings saves.
-    if ((local.inputTextProvided || local.deviceModeProvided) &&
-        clusterFollowerViewGet().gated) {
-      if (isAjax) request->send(409, "text/plain", F("clustered"));
-      else request->redirect("/?clustered=true");
-      return;
-    }
-    if ((local.inputTextProvided || local.transientTextProvided) &&
-        displayQueueFull()) {
-      SerialPrintln(F("Display command queue full — message rejected."));
-      if (isAjax) request->send(503, "text/plain", F("display busy"));
-      else request->redirect("/?display-busy=true");
-      return;
-    }
-
-    // Verdict against the live values, then stage; the apply itself runs in
-    // webEndpointsLoop(). Verdict + merge sit in one locked section so the
-    // comparison can't race a half-applied post.
-    bool needsReboot;
-    bool deviceNameChanged;
-    {
-      WebStateLock lock;
-      needsReboot = settingsPostNeedsReboot(local, *liveSettings);
-      deviceNameChanged = local.deviceNameProvided &&
-                          local.deviceName != liveSettings->deviceName;
-      mergeSettingsPost(pendingPost, local);
-    }
-
-    // Device renamed (#125): flag only from async context — mqttTask blanks
-    // the old identity's retained discovery configs before the reboot swaps
-    // identities.
-    if (deviceNameChanged) mqttRequestDiscoveryClear();
-
-    if (isAjax) {
-      request->send(200, "text/plain", needsReboot ? F("ok-reboot") : F("ok"));
-    } else {
-      request->redirect(deviceNameChanged ? "/?device-name-saved=true"
-                        : needsReboot     ? "/?mqtt-saved=true"
-                                          : "/");
-    }
-  });
-
-  // POST, not GET (v1 #145): state-changing actions must not be triggerable
-  // by a drive-by <img src> on the LAN.
-  server.on("/reboot", HTTP_POST, [](AsyncWebServerRequest* request) {
-    SerialPrintln(F("Reboot requested from web UI"));
-    request->send(200, "text/plain",
-                  "Reboot pending — this takes a few seconds. Reload the home "
-                  "page afterwards.");
-    WebStateLock lock;
-    pendingReboot = true;
-    rebootRequestedAtMs = millis();
-  });
-
-  // --- WiFi portal + credentials (#188) -------------------------------------
-  // Handlers stage into WifiService; all radio/NVS work runs in netTask's
-  // wifiServiceTick(). The portal page itself is served unconditionally —
-  // from the LAN it doubles as a "move to another network" page.
-  server.on("/wifi-setup", HTTP_GET, [](AsyncWebServerRequest* request) {
-    serveGzipAsset(request, "text/html", PORTAL_HTML_GZ, PORTAL_HTML_GZ_LEN);
-  });
-
-  server.on("/wifi/scan", HTTP_POST, [](AsyncWebServerRequest* request) {
-    wifiStageScan();
-    request->send(200, "text/plain", F("scanning"));
-  });
-  server.on("/wifi/scan", HTTP_GET, [](AsyncWebServerRequest* request) {
-    String json = wifiScanResultJson();
-    if (json.length() == 0) {
-      request->send(202, "text/plain", F("pending"));
-    } else {
-      request->send(200, "application/json", json);
-    }
-  });
-
-  server.on("/wifi/config", HTTP_POST, [](AsyncWebServerRequest* request) {
-    SerialPrintln(F("WiFi credentials submitted from web"));
-    if (!request->hasParam("ssid", true)) {
-      request->send(400, "text/plain", F("invalid"));
-      return;
-    }
-    String ssid = request->getParam("ssid", true)->value();
-    String pass = request->hasParam("pass", true)
-                      ? request->getParam("pass", true)->value()
-                      : String();
-    if (!isValidWifiSsidValue(ssid, LEN_WIFI_SSID) ||
-        !isValidWifiPasswordValue(pass, LEN_WIFI_PASSWORD)) {
-      SerialPrintln(F("WiFi config rejected: invalid ssid/password"));
-      request->send(400, "text/plain", F("invalid"));
-      return;
-    }
-    wifiStagePortalConfig(ssid, pass);
-    request->send(200, "text/plain", F("ok-reboot"));
-  });
-
-  // POST, not GET (v1 #145): erasing WiFi credentials is the sharpest edge.
-  server.on("/reset-wifi", HTTP_POST, [](AsyncWebServerRequest* request) {
-    SerialPrintln(F("WiFi credential erase requested from web UI"));
-    request->send(200, "text/plain",
-                  "WiFi credentials erased. The display is rebooting into its "
-                  "setup portal — reconnect to the device's setup AP to "
-                  "configure a network.");
-    wifiStageReset();
-  });
-
-  // Captive-portal hook: while the setup portal is up, the DNS catch-all
-  // funnels every hostname here and this redirect pops the OS sign-in sheet
-  // (/generate_204, /hotspot-detect.html, /connecttest.txt all land in
-  // onNotFound). Outside portal mode: a plain 404, as v1.
-  server.onNotFound([](AsyncWebServerRequest* request) {
-    String redirectUrl = wifiPortalRedirectUrl();  // "" = portal not up
-    if (redirectUrl.length() > 0) {
-      request->redirect(redirectUrl);
-    } else {
-      request->send(404, "text/plain", F("Not found"));
-    }
-  });
-
-  // --- master OTA (#190) -----------------------------------------------------
-  // v1 wire contract: POST multipart field "firmware" + mandatory ?md5=
-  // (v1 #144) + optional ?v= intended-version diagnostic. Update targets
-  // the inactive A/B slot; the image boots PENDING_VERIFY and OtaService
-  // confirms it pre-inrush near the end of setup() (#305), netif-up as
-  // fallback (bootloader reverts otherwise).
-  //
-  // Async-context exception (deliberate, v1 precedent): Update.write() runs
-  // right here in the async_tcp task — a firmware stream cannot be staged
-  // through a queue. The handler still never touches settings/NVS/radio;
-  // the ?v= write and the reboot are staged for the netTask drain.
-  server.on(
-      "/firmware/master", HTTP_POST,
-      [](AsyncWebServerRequest* request) {
-        // Overlap-rejected upload (#191), or a bodyless POST racing a live
-        // session: answer 409 without touching the owner's shared state,
-        // MQTT freeze or Update session.
-        if (request->_tempObject != nullptr ||
-            (otaOwnerRequest != nullptr && otaOwnerRequest != request)) {
-          request->send(409, "text/plain",
-                        F("Another master OTA upload is already in progress "
-                          "— retry when it finishes"));
-          return;
-        }
-        otaOwnerRequest = nullptr;  // session concluded, whatever the verdict
-        if (otaRejectionStatus != 0) {
-          request->send(otaRejectionStatus, "text/plain", otaRejectionReason);
-          return;
-        }
-        if (Update.hasError()) {
-          mqttResumeAfterOta();  // no reboot coming — thaw the session (#116)
-          request->send(500, "text/plain", String("Master OTA failed: ") +
-                                               Update.errorString());
-          return;
-        }
-        if (!Update.isFinished()) {
-          mqttResumeAfterOta();  // no reboot coming — thaw the session (#116)
-          request->send(500, "text/plain",
-                        F("Master OTA incomplete: upload ended before the "
-                          "image was complete."));
-          return;
-        }
-        request->send(200, "text/plain",
-                      F("Master firmware flashed; rebooting…"));
-        WebStateLock lock;
-        pendingReboot = true;
-        rebootRequestedAtMs = millis();
-      },
-      [](AsyncWebServerRequest* request, String filename, size_t index,
-         uint8_t* data, size_t len, bool final) {
-        if (index == 0) {
-          // Concurrent-upload guard (#191): a live session owns the Update
-          // singleton and the shared rejection state — mark this request
-          // rejected (per-request _tempObject; malloc pairs with the free
-          // in the request destructor) and leave both alone.
-          if (otaOwnerRequest != nullptr && otaOwnerRequest != request) {
-            request->_tempObject = malloc(1);
-            return;
-          }
-          otaRejectionStatus = 0;
-          otaRejectionReason = "";
-
-          // CSRF gate (#313), INLINE before Update.begin: the middleware runs
-          // too late for upload routes (post-body), so a forged cross-site
-          // POST would otherwise flash + arm its image before the 403.
-          if (webUploadCsrfRejected(request)) {
-            otaRejectionStatus = 403;
-            otaRejectionReason = "Cross-origin OTA refused (CSRF guard)";
-            return;
-          }
-
-          // Reflash gate (#205): a master OTA reboots the S3 mid-unit-flash
-          // and strands the in-flight unit in twiboot for no reason.
-          if (reflashInProgress(displaySnapshotGet().reflash)) {
-            otaRejectionStatus = 409;
-            otaRejectionReason =
-                "Unit reflash in progress — retry when it finishes";
-            return;
-          }
-
-          String md5 = request->hasParam("md5")
-                           ? request->getParam("md5")->value()
-                           : String();
-          if (md5.length() == 0) {
-            otaRejectionStatus = 400;
-            otaRejectionReason =
-                "md5 query parameter is required (compute it over the .bin "
-                "and pass ?md5=...)";
-            return;
-          }
-          if (!normalizeOtaMd5(md5)) {
-            otaRejectionStatus = 400;
-            otaRejectionReason = "md5 must be exactly 32 hex characters";
-            return;
-          }
-
-          SerialPrintln("Master OTA upload started (md5 " + md5 + ")");
-          if (Update.isRunning()) {
-            Update.abort();  // stale aborted upload must not wedge this one
-                             // (v1 #162 re-entry class)
-          }
-          if (!Update.begin(UPDATE_SIZE_UNKNOWN, U_FLASH)) {
-            otaRejectionStatus = 500;
-            otaRejectionReason =
-                String("Master OTA could not start: ") + Update.errorString();
-            return;
-          }
-          Update.setMD5(md5.c_str());
-          // Freeze MQTT for the flash writes (v1 #116): staged — mqttTask
-          // force-closes the session, the broker fires the "offline" LWT.
-          mqttStopForOta();
-
-          // ?v= staged unconditionally (empty if absent) so a stale value
-          // from an earlier flash can't outlive this one (v1 #52 rationale),
-          // through the same validator as every other settings write (#191).
-          String intended = sanitizeIntendedVersion(
-              request->hasParam("v") ? request->getParam("v")->value()
-                                     : String());
-          {
-            WebStateLock lock;
-            pendingIntendedVersion = intended;
-            pendingIntendedVersionProvided = true;
-          }
-
-          // Session is live from here (#191). onDisconnect is the backstop
-          // for a client that dies mid-upload: free the slot; the stale
-          // Update session is aborted by the next upload's begin path above.
-          otaOwnerRequest = request;
-          otaUploadStartMs = millis();
-          otaLastChunkMs = otaUploadStartMs;  // #313 stall watchdog baseline
-          request->onDisconnect([request]() {
-            if (otaOwnerRequest == request) otaOwnerRequest = nullptr;
-          });
-        }
-
-        // Not (or no longer) the live owner (#191): covers overlap-rejected
-        // requests AND stragglers rejected via the shared flag above whose
-        // client keeps streaming — a later legitimate owner resets that
-        // flag, and without this gate their leftover chunks would write
-        // into the new owner's session.
-        if (otaOwnerRequest != request) return;
-        if (otaRejectionStatus != 0) return;
-        otaLastChunkMs = millis();  // #313: progress resets the stall deadline
-
-        if (len > 0 && Update.write(data, len) != len) {
-          // Error is latched inside Update; the completion callback
-          // reports it. Stop consuming flash time on further chunks.
-          return;
-        }
-        if (final) {
-          if (Update.end(true)) {
-            uint32_t elapsedMs = millis() - otaUploadStartMs;
-            uint32_t totalBytes = (uint32_t)(index + len);
-            if (elapsedMs == 0) elapsedMs = 1;
-            SerialPrintf(
-                "Master OTA received %u bytes in %u ms (%u KB/s)\n",
-                (unsigned)totalBytes, (unsigned)elapsedMs,
-                (unsigned)(((uint64_t)totalBytes * 1000ULL / 1024ULL) /
-                           elapsedMs));
-            SerialPrintln(F("Master OTA image verified and armed — reboot "
-                            "boots it PENDING_VERIFY"));
-          } else {
-            SerialPrintln(String("Master OTA failed at end: ") +
-                          Update.errorString());
-          }
-        }
-      });
-
-  server.on("/debug/ota", HTTP_GET, [](AsyncWebServerRequest* request) {
-    request->send(200, "application/json", otaDebugJson());
-  });
-
-  // --- factory rescue slot (#195) --------------------------------------------
-  // Install/refresh the rescue image over WiFi: same wire contract as
-  // /firmware/master (multipart field "firmware" + mandatory ?md5=), target
-  // = the factory partition via FactorySlot's raw esp_partition writes.
-  // Never touches otadata — installing a rescue image does not change what
-  // boots next. Same async-context exception as the master OTA above: the
-  // firmware stream is written right here in the async_tcp task.
-  server.on(
-      "/firmware/rescue", HTTP_POST,
-      [](AsyncWebServerRequest* request) {
-        if (rescueRejectionStatus != 0) {
-          request->send(rescueRejectionStatus, "text/plain",
-                        rescueRejectionReason);
-          return;
-        }
-        String err = factoryWriteError();
-        if (err.length() > 0) {
-          request->send(500, "text/plain", "Rescue install failed: " + err);
-          return;
-        }
-        request->send(200, "text/plain",
-                      F("Rescue image installed into the factory slot. No "
-                        "reboot — POST /firmware/rescue-boot to test it."));
-      },
-      [](AsyncWebServerRequest* request, String filename, size_t index,
-         uint8_t* data, size_t len, bool final) {
-        if (index == 0) {
-          // An install is already streaming: 409 without touching its state.
-          // (The shared rejection flag stalls the in-flight upload too — the
-          // erased header keeps that safe; clean per-request verdicts are
-          // #191 territory.)
-          if (factoryInstallInProgress()) {
-            rescueRejectionStatus = 409;
-            rescueRejectionReason =
-                "another rescue install is in flight — let it finish (a "
-                "dropped one expires after ~30 s) and retry";
-            return;
-          }
-          rescueRejectionStatus = 0;
-          rescueRejectionReason = "";
-
-          // CSRF gate (#313), INLINE before factoryWriteBegin — the
-          // middleware fires post-body, too late for an upload route.
-          if (webUploadCsrfRejected(request)) {
-            rescueRejectionStatus = 403;
-            rescueRejectionReason = "Cross-origin rescue upload refused (CSRF "
-                                    "guard)";
-            return;
-          }
-
-          String md5 = request->hasParam("md5")
-                           ? request->getParam("md5")->value()
-                           : String();
-          if (md5.length() == 0) {
-            rescueRejectionStatus = 400;
-            rescueRejectionReason =
-                "md5 query parameter is required (compute it over the .bin "
-                "and pass ?md5=...)";
-            return;
-          }
-          if (!normalizeOtaMd5(md5)) {
-            rescueRejectionStatus = 400;
-            rescueRejectionReason = "md5 must be exactly 32 hex characters";
-            return;
-          }
-          if (!factoryWriteBegin(md5)) {
-            rescueRejectionStatus = 500;
-            rescueRejectionReason =
-                "Rescue install could not start: " + factoryWriteError();
-            return;
-          }
-        }
-
-        if (rescueRejectionStatus != 0) return;
-
-        if (len > 0 && !factoryWriteChunk(data, len, index)) {
-          return;  // error latched in FactorySlot; completion reports it
-        }
-        if (final) {
-          factoryWriteEnd();  // completion callback reads factoryWriteError()
-        }
-      });
-
-  // Software entry into the rescue image — and the periodic "prove the
-  // rescue image still boots" test (#195 spec). Guarded so a wall-mounted
-  // device can't be pointed at an empty slot (the bootloader would fall
-  // back to an OTA slot anyway, but the endpoint must stay honest as a
-  // boot test). otadata erase + staged reboot; NVS untouched, so WiFi
-  // credentials survive into rescue (#193 invariant).
-  server.on("/firmware/rescue-boot", HTTP_POST,
-            [](AsyncWebServerRequest* request) {
-              if (factoryInstallInProgress()) {
-                request->send(409, "text/plain",
-                              F("A rescue image install is in flight — let "
-                                "it finish, then retry."));
-                return;
-              }
-              if (!factorySlotImageValid()) {
-                request->send(409, "text/plain",
-                              F("Factory slot holds no valid rescue image — "
-                                "POST it to /firmware/rescue first."));
-                return;
-              }
-              if (!rescueBootArm()) {
-                request->send(500, "text/plain",
-                              F("otadata erase failed — rescue boot not "
-                                "armed."));
-                return;
-              }
-              request->send(200, "text/plain",
-                            F("Rebooting into the rescue image… it joins "
-                              "WiFi (or opens <name>-rescue) and serves the "
-                              "recovery page."));
-              WebStateLock lock;
-              pendingReboot = true;
-              rebootRequestedAtMs = millis();
-            });
-
-  // --- unit health (#203, v1 #45 wire contract) -----------------------------
-  // GET renders JSON from the snapshot copy — never touches the bus from
-  // async context (the async rule is structural here: only displayTask
-  // holds Wire).
-  // System tab (#245): current vitals + ~10 min sparkline history in one
-  // JSON. History is server-side (netTask's sample ring) so a freshly
-  // opened tab has depth immediately; the browser polls at 2 s.
-  server.on("/system/stats", HTTP_GET, [](AsyncWebServerRequest* request) {
-    std::unique_ptr<char[]> buf(new char[SYSTEM_STATS_JSON_CAP]);
-    size_t n = systemStatsJson(buf.get(), SYSTEM_STATS_JSON_CAP);
-    if (n == 0 || n >= SYSTEM_STATS_JSON_CAP) {
-      request->send(500, "text/plain", F("stats unavailable"));
-      return;
-    }
-    request->send(200, "application/json", buf.get());
-  });
-
-  // Self-documenting route + terse-key legend index for the headless
-  // (curl-only) operator (#307). Static data, heap-rendered per request.
-  server.on("/api", HTTP_GET, [](AsyncWebServerRequest* request) {
-    std::unique_ptr<char[]> buf(new char[API_JSON_CAP]);
-    size_t n = buildApiJson(buf.get(), API_JSON_CAP);
-    if (n == 0 || n >= API_JSON_CAP) {
-      request->send(500, "text/plain", F("api index unavailable"));
-      return;
-    }
-    request->send(200, "application/json", buf.get());
-  });
-
-  // Static hardware/partition inventory (#307). Reuses existing accessors; no
-  // new sampling. bootCount comes from the NVS counter main.cpp bumps.
-  server.on("/system/info", HTTP_GET, [](AsyncWebServerRequest* request) {
-    String out;
-    out.reserve(1024);
-    out += "{\"chip\":\"";
-    out += ESP.getChipModel();
-    out += "\",\"chipRev\":";
-    out += String(ESP.getChipRevision());
-    out += ",\"cores\":";
-    out += String(ESP.getChipCores());
-    out += ",\"cpuMHz\":";
-    out += String(ESP.getCpuFreqMHz());
-    out += ",\"flashKB\":";
-    out += String(ESP.getFlashChipSize() / 1024);
-    out += ",\"psramKB\":";
-    out += String(ESP.getPsramSize() / 1024);
-    out += ",\"rev\":\"";
-    out += GIT_REV;
-    out += "\",\"bundledUnitRev\":\"";
-    out += BUNDLED_UNIT_REV;
-    out += "\",\"sketchMd5\":\"";
-    out += ESP.getSketchMD5();
-    out += "\",\"bootCount\":";
-    out += String(liveStore != nullptr ? liveStore->getInt("bootCount", 0) : 0);
-    out += ",\"resetReason\":\"";
-    out += webResetReasonString();
-    out += "\",\"factoryPresent\":";
-    out += factorySlotPresent() ? "true" : "false";
-    out += ",\"rescueValid\":";  // factory slot holds a bootable rescue image
-    out += factorySlotImageValid() ? "true" : "false";
-    out += ",\"partitions\":[";
-    esp_partition_iterator_t it = esp_partition_find(
-        ESP_PARTITION_TYPE_ANY, ESP_PARTITION_SUBTYPE_ANY, nullptr);
-    bool first = true;
-    while (it != nullptr) {
-      const esp_partition_t* p = esp_partition_get(it);
-      if (!first) out += ',';
-      first = false;
-      out += "{\"label\":\"";
-      out += p->label;
-      out += "\",\"type\":";
-      out += String(p->type);
-      out += ",\"subtype\":";
-      out += String(p->subtype);
-      out += ",\"offset\":";
-      out += String((unsigned long)p->address);
-      out += ",\"size\":";
-      out += String((unsigned long)p->size);
-      out += '}';
-      it = esp_partition_next(it);
-    }
-    esp_partition_iterator_release(it);
-    out += "]}";
-    request->send(200, "application/json", out);
-  });
-
-  // One-shot aggregate for a single curl (#307): settings + stats.now + units
-  // + cluster + ota, composed from the existing serializers. History stays at
-  // /system/stats to keep this bounded.
-  server.on("/status", HTTP_GET, [](AsyncWebServerRequest* request) {
-    std::unique_ptr<char[]> nowBuf(new char[SYSTEM_STATS_JSON_CAP]);
-    size_t nowN = systemStatsNowJson(nowBuf.get(), SYSTEM_STATS_JSON_CAP);
-    if (nowN >= SYSTEM_STATS_JSON_CAP) nowBuf[0] = '\0';
-
-    DisplaySnapshot snap = displaySnapshotGet();
-    std::unique_ptr<char[]> unitsBuf(new char[UNIT_HEALTH_JSON_CAP]);
-    size_t unitsN =
-        buildUnitHealthJson(unitsBuf.get(), UNIT_HEALTH_JSON_CAP, snap.units,
-                            snap.displayWidth, snap.faultyUnitCount,
-                            SFP_I2C_ADDRESS_BASE, millis());
-    if (unitsN == 0 || unitsN >= UNIT_HEALTH_JSON_CAP) {
-      snprintf(unitsBuf.get(), UNIT_HEALTH_JSON_CAP,
-               "{\"width\":%d,\"faulty\":%d,\"units\":[]}", snap.displayWidth,
-               snap.faultyUnitCount);
-    }
-
-    String out;
-    out.reserve(6144);
-    out += "{\"settings\":";
-    out += buildCurrentSettingsJson();
-    out += ",\"stats\":{\"now\":";
-    out += nowBuf.get();
-    out += "},\"units\":";
-    out += unitsBuf.get();
-    out += ",\"cluster\":";
-    out += clusterStatusJson(clusterLeaderStatusGet());
-    out += ",\"ota\":";
-    out += otaDebugJson();
-    out += "}";
-    request->send(200, "application/json", out);
-  });
-
-  server.on("/units/health", HTTP_GET, [](AsyncWebServerRequest* request) {
-    DisplaySnapshot snap = displaySnapshotGet();
-    // Heap, not stack: ~2 KB doesn't belong on the async_tcp task stack,
-    // and a static buffer would race concurrent requests.
-    std::unique_ptr<char[]> buf(new char[UNIT_HEALTH_JSON_CAP]);
-    size_t n =
-        buildUnitHealthJson(buf.get(), UNIT_HEALTH_JSON_CAP, snap.units,
-                            snap.displayWidth, snap.faultyUnitCount,
-                            SFP_I2C_ADDRESS_BASE, millis());
-    if (n == 0 || n >= UNIT_HEALTH_JSON_CAP) {
-      // Would-be-truncated payload: fall back to a valid headline-only JSON
-      // rather than shipping a cut object (v1 truncation discipline).
-      n = (size_t)snprintf(buf.get(), UNIT_HEALTH_JSON_CAP,
-                           "{\"width\":%d,\"faulty\":%d,\"units\":[]}",
-                           snap.displayWidth, snap.faultyUnitCount);
-    }
-    // Wear assessment rides the same payload (#231, additive key), spliced
-    // before the closing brace like the reflash object below.
-    WearAssessment wear;
-    assessWear(snap.units, snap.displayWidth, wear);
-    char wearJson[96];
-    size_t wearLen = buildWearJson(wear, wearJson, sizeof(wearJson));
-    if (n > 0 && wearLen < sizeof(wearJson) &&
-        n + wearLen + 2 < UNIT_HEALTH_JSON_CAP) {
-      n += (size_t)snprintf(buf.get() + n - 1, UNIT_HEALTH_JSON_CAP - n + 1,
-                            ",%s}", wearJson) - 1;
-    }
-    // Reflash progress rides the same payload (#205, additive key — the
-    // Maintenance tab already polls this endpoint). Spliced before the
-    // closing brace; the ~70 B worst case fits the cap's slack by design.
-    char reflashJson[80];
-    buildReflashJson(reflashJson, sizeof(reflashJson), snap.reflash);
-    if (n > 0 && n + strlen(reflashJson) + 13 < UNIT_HEALTH_JSON_CAP) {
-      snprintf(buf.get() + n - 1, UNIT_HEALTH_JSON_CAP - n + 1,
-               ",\"reflash\":%s}", reflashJson);
-    }
-    request->send(200, "application/json", buf.get());
-  });
-  // POST enqueues a Probe — displayTask re-scans the bus AND re-polls
-  // health in one pass (a probe subsumes v1's plain re-poll; ?probe=1 is
-  // accepted for wire compat but changes nothing). 202 mirrors v1's
-  // "pending"; a full queue reports busy like v1's flash-in-progress 503.
-  server.on("/units/health/refresh", HTTP_POST,
-            [](AsyncWebServerRequest* request) {
-              if (rejectWhileReflashing(request)) return;
-              if (!displayEnqueue(makeProbeCommand())) {
-                request->send(503, "application/json",
-                              F("{\"status\":\"busy\"}"));
-                return;
-              }
-              request->send(202, "application/json",
-                            F("{\"status\":\"pending\"}"));
-            });
-
-  // --- calibration + provisioning (#204, queue-native) ----------------------
-  // Handlers validate against a snapshot copy, enqueue with a fresh seq and
-  // answer {"seq":N}; execution truth arrives via GET /unit/op-result plus
-  // the refreshed health facts. The async rule is upheld throughout: no
-  // handler touches the bus or waits on display work.
-
-  server.on("/unit/offset", HTTP_GET, [](AsyncWebServerRequest* request) {
-    DisplaySnapshot snap = displaySnapshotGet();
-    int addr = 0;
-    if (!maintCheckAddress(request, snap, addr)) return;
-    const UnitFacts& unit = snap.units[addr - SFP_I2C_ADDRESS_BASE];
-    if (!unit.offsetValid) {
-      // v1 wording (pre-#32 unit firmware has no GET_OFFSET); also covers
-      // "reads invalidated until the next probe".
-      request->send(502, "text/plain",
-                    F("Unit did not return a valid offset (firmware may "
-                      "predate #32)"));
-      return;
-    }
-    char buf[32];
-    snprintf(buf, sizeof(buf), "{\"offset\":%d}", (int)unit.offset);
-    request->send(200, "application/json", buf);
-  });
-
-  server.on("/unit/offset", HTTP_POST, [](AsyncWebServerRequest* request) {
-    DisplaySnapshot snap = displaySnapshotGet();
-    int addr = 0;
-    if (!maintCheckAddress(request, snap, addr)) return;
-    long value = 0;
-    if (!maintRequireLongParam(request, "value", value)) return;
-    MaintVerdict verdict = maintValidateOffset(value);
-    if (verdict.httpStatus != 200) {
-      request->send(verdict.httpStatus, "text/plain", verdict.message);
-      return;
-    }
-    maintEnqueue(request, makeWriteOffsetCommand(displayNextMaintSeq(),
-                                                 (uint8_t)addr,
-                                                 (int16_t)value));
-  });
-
-  server.on("/unit/jog", HTTP_POST, [](AsyncWebServerRequest* request) {
-    DisplaySnapshot snap = displaySnapshotGet();
-    int addr = 0;
-    if (!maintCheckAddress(request, snap, addr)) return;
-    long steps = 0;
-    if (!maintRequireLongParam(request, "steps", steps)) return;
-    MaintVerdict verdict = maintValidateJog(steps);
-    if (verdict.httpStatus != 200) {
-      request->send(verdict.httpStatus, "text/plain", verdict.message);
-      return;
-    }
-    maintEnqueue(request, makeJogCommand(displayNextMaintSeq(), (uint8_t)addr,
-                                         (int)steps));
-  });
-
-  server.on("/unit/home", HTTP_POST, [](AsyncWebServerRequest* request) {
-    DisplaySnapshot snap = displaySnapshotGet();
-    int addr = 0;
-    if (!maintCheckAddress(request, snap, addr)) return;
-    maintEnqueue(request,
-                 makeHomeCommand(displayNextMaintSeq(), (uint8_t)addr));
-  });
-
-  server.on("/unit/identify", HTTP_POST, [](AsyncWebServerRequest* request) {
-    DisplaySnapshot snap = displaySnapshotGet();
-    int addr = 0;
-    if (!maintCheckAddress(request, snap, addr)) return;
-    maintEnqueue(request,
-                 makeIdentifyCommand(displayNextMaintSeq(), (uint8_t)addr));
-  });
-
-  // Physical-rebuild bookkeeping (#231): zero the unit's wear odometer after
-  // a flap swap or motor replacement. Same op contract as identify/home.
-  server.on("/unit/reset-odometer", HTTP_POST,
-            [](AsyncWebServerRequest* request) {
-    DisplaySnapshot snap = displaySnapshotGet();
-    int addr = 0;
-    if (!maintCheckAddress(request, snap, addr)) return;
-    maintEnqueue(request, makeResetOdometerCommand(displayNextMaintSeq(),
-                                                   (uint8_t)addr));
-  });
-
-  // On-demand unit self-test (#265): the unit measures its own mechanics
-  // (steps/rev, hall window, rev time) over ~2 revolutions. Same op
-  // contract as identify/home; the measurements come back via
-  // GET /unit/self-test-result.
-  server.on("/unit/self-test", HTTP_POST, [](AsyncWebServerRequest* request) {
-    DisplaySnapshot snap = displaySnapshotGet();
-    int addr = 0;
-    if (!maintCheckAddress(request, snap, addr)) return;
-    maintEnqueue(request,
-                 makeSelfTestCommand(displayNextMaintSeq(), (uint8_t)addr));
-  });
-
-  // Self-test execution feedback: pending / ok(+measurements) /
-  // failed(+reason) / expired, rendered from the snapshot's single
-  // SelfTestSlot — the self-test twin of /unit/op-result.
-  server.on("/unit/self-test-result", HTTP_GET,
-            [](AsyncWebServerRequest* request) {
-    long seq = 0;
-    if (!maintRequireLongParam(request, "seq", seq)) return;
-    if (seq < 1) {
-      request->send(400, "text/plain", F("seq must be >= 1"));
-      return;
-    }
-    DisplaySnapshot snap = displaySnapshotGet();
-    char buf[128];
-    buildSelfTestJson(buf, sizeof(buf), snap.lastSelfTest, (uint32_t)seq);
-    request->send(200, "application/json", buf);
-  });
-
-  // Debug endpoint, v1 semantics preserved: pushes the unit into twiboot
-  // (~1 s on its DIP-derived address, then back to the sketch). v1 parity:
-  // range check only, no sketch-state gate — it exists precisely for poking
-  // at units the probe view might mislabel. displayTask deliberately does
-  // NOT reprobe afterwards (v1 #88: probing the twiboot window pins the
-  // bootloader alive).
-  server.on("/unit/reboot", HTTP_POST, [](AsyncWebServerRequest* request) {
-    long addr = 0;
-    if (!maintRequireLongParam(request, "address", addr)) return;
-    if (addr < 1 || addr > 126) {
-      request->send(400, "text/plain", F("Address must be 1..126"));
-      return;
-    }
-    maintEnqueue(request, makeRebootToBootloaderCommand(displayNextMaintSeq(),
-                                                        (uint8_t)addr));
-  });
-
-  server.on("/unit/set-address", HTTP_POST,
-            [](AsyncWebServerRequest* request) {
-              DisplaySnapshot snap = displaySnapshotGet();
-              int addr = 0;
-              if (!maintCheckAddress(request, snap, addr)) return;
-              long target = 0;
-              if (!maintRequireLongParam(request, "value", target)) return;
-              // Fast 409 from the copy; displayTask re-runs the same policy
-              // against live facts right before the burn (authoritative).
-              MaintVerdict verdict = maintValidateSetAddressTarget(
-                  target, addr, snap.units, UNITS_AMOUNT);
-              if (verdict.httpStatus != 200) {
-                request->send(verdict.httpStatus, "text/plain",
-                              verdict.message);
-                return;
-              }
-              maintEnqueue(request,
-                           makeSetAddressCommand(displayNextMaintSeq(),
-                                                 (uint8_t)addr,
-                                                 (uint8_t)target));
-            });
-
-  server.on("/unit/clear-address", HTTP_POST,
-            [](AsyncWebServerRequest* request) {
-              DisplaySnapshot snap = displaySnapshotGet();
-              int addr = 0;
-              if (!maintCheckAddress(request, snap, addr)) return;
-              maintEnqueue(request, makeClearAddressCommand(
-                                        displayNextMaintSeq(), (uint8_t)addr));
-            });
-
-  // Blank-out recalibration (v1 semantics: two full-row frames force the
-  // wrap-around re-home, then the enqueue-time text returns). Text comes
-  // from the display snapshot (what is actually showing), alignment/speed
-  // from the settings of this moment — all baked into the command.
-  server.on("/reset-units", HTTP_POST, [](AsyncWebServerRequest* request) {
-    SerialPrintln(F("Units reset requested from web UI"));
-    DisplaySnapshot snap = displaySnapshotGet();
-    WebContentSnapshot content = webDisplayContentSnapshot();
-    maintEnqueue(request, makeResetUnitsCommand(
-                              displayNextMaintSeq(), String(snap.currentText),
-                              content.alignment, content.flapSpeed));
-  });
-
-  // Kill switch (v1 #35). Order is load-bearing (review 2026-07-11): the
-  // abort flag is set BEFORE the enqueue so the queue's happens-before
-  // guarantees displayTask's Stop always finds it set — set-after-enqueue
-  // races an idle displayTask clearing it first, stranding the flag ON for
-  // every future wait. A 503 rolls the flag back (nothing queued to abort).
-  server.on("/stop", HTTP_POST, [](AsyncWebServerRequest* request) {
-    SerialPrintln(F("Stop requested from web UI"));
-    DisplayCommand cmd = makeStopCommand(displayNextMaintSeq());
-    unitBusRequestAbort();
-    if (!displayEnqueue(cmd)) {
-      unitBusClearAbort();
-      request->send(503, "text/plain",
-                    F("Display queue full — try again in a moment"));
-      return;
-    }
-    // Cluster propagation (#317): when leading, Stop blanks the WHOLE wall —
-    // the local command above handles this board's own row; blank the
-    // followers in sync (no-op when standalone).
-    clusterLeaderBlankWall();
-    char buf[24];
-    snprintf(buf, sizeof(buf), "{\"seq\":%lu}", (unsigned long)cmd.seq);
-    request->send(200, "application/json", buf);
-  });
-
-  // Execution feedback for the queued ops: pending / ok / failed(+reason) /
-  // expired, rendered from the snapshot's single MaintResult slot.
-  server.on("/unit/op-result", HTTP_GET, [](AsyncWebServerRequest* request) {
-    long seq = 0;
-    if (!maintRequireLongParam(request, "seq", seq)) return;
-    if (seq < 1) {
-      request->send(400, "text/plain", F("seq must be >= 1"));
-      return;
-    }
-    DisplaySnapshot snap = displaySnapshotGet();
-    char buf[96];
-    buildOpResultJson(buf, sizeof(buf), snap.lastMaint, (uint32_t)seq);
-    request->send(200, "application/json", buf);
-  });
-
-  // Bulk unit reflash (#205): pushes every unit not on the bundled rev
-  // through twiboot with the PROGMEM-embedded image. Queue-native like all
-  // maintenance ops — {"seq":N} now, job outcome via /unit/op-result, live
-  // progress in /units/health's reflash object. Text/alignment/speed baked
-  // at enqueue (the job re-shows them; reflashed units homed to blank).
-  server.on("/reflash-units", HTTP_POST, [](AsyncWebServerRequest* request) {
-    SerialPrintln(F("Unit reflash requested from web UI"));
-    DisplaySnapshot snap = displaySnapshotGet();
-    WebContentSnapshot content = webDisplayContentSnapshot();
-    maintEnqueue(request, makeReflashUnitsCommand(
-                              displayNextMaintSeq(), String(snap.currentText),
-                              content.alignment, content.flapSpeed));
-  });
-
-  // --- MQTT broker mDNS discovery (#224, v1 /mqtt/discover contract) --------
-  // POST arms the staged flag (re-POST while pending is a no-op — the flag
-  // is the re-entry guard); the blocking MDNS.queryService pass runs from
-  // netTask's drain, never here.
-  server.on("/mqtt/discover", HTTP_POST, [](AsyncWebServerRequest* request) {
-    {
-      WebStateLock lock;
-      if (!mqttDiscoverPending) {
-        mqttDiscoverResultJson = "";
-        mqttDiscoverPending = true;
-      }
-    }
-    request->send(200, "text/plain", F("Broker discovery started"));
-  });
-  server.on("/mqtt/discover", HTTP_GET, [](AsyncWebServerRequest* request) {
-    // Snapshot under the lock, send outside it (file convention — send()
-    // schedules through AsyncTCP and its lock domain must not nest ours).
-    bool pending;
-    String json;
-    {
-      WebStateLock lock;
-      pending = mqttDiscoverPending;
-      json = mqttDiscoverResultJson;
-    }
-    if (pending) {
-      request->send(202, "text/plain", F("Discovery running"));
-      return;
-    }
-    if (json.length() == 0) {
-      request->send(404, "text/plain", F("No discovery has run yet"));
-      return;
-    }
-    request->send(200, "application/json", json);
-  });
-
-  // --- Cluster follower endpoints (#272, epic #270) --------------------------
-  // The LAN wire protocol the leader drives (form-encoded requests, JSON
-  // replies — same conventions as the rest of this API). Handlers stage
-  // into ClusterFollower under its own mutex; NVS writes and the display
-  // enqueue run in netTask's clusterFollowerServiceTick().
-  server.on("/cluster/join", HTTP_POST, [](AsyncWebServerRequest* request) {
-    if (!request->hasParam("leaderHost", true) ||
-        !request->hasParam("row", true) || !request->hasParam("epoch", true)) {
-      request->send(400, "text/plain", F("Missing leaderHost/row/epoch"));
-      return;
-    }
-    ClusterJoinRequest req;
-    req.leaderHost = request->getParam("leaderHost", true)->value();
-    req.leaderName = request->hasParam("leaderName", true)
-                         ? request->getParam("leaderName", true)->value()
-                         : req.leaderHost;
-    // #313 follow-on: the leader mints a per-member wire-auth key and sends it
-    // here (absent from a pre-HMAC leader → enforcement stays off).
-    req.key = request->hasParam("key", true)
-                  ? request->getParam("key", true)->value()
-                  : String();
-    long row = request->getParam("row", true)->value().toInt();
-    req.epoch = (uint32_t)strtoul(
-        request->getParam("epoch", true)->value().c_str(), nullptr, 10);
-    if (row < 0 || row >= CLUSTER_MAX_MEMBERS) {
-      request->send(400, "text/plain", F("Row out of range"));
-      return;
-    }
-    req.row = (int)row;
-    if (req.leaderHost.length() == 0 ||
-        req.leaderHost.length() > CLUSTER_HOST_MAX_LEN ||
-        !settingsIsPrintableAscii(req.leaderHost, 0x21)) {
-      request->send(400, "text/plain", F("Invalid leaderHost"));
-      return;
-    }
-    if (req.leaderName.length() > CLUSTER_HOST_MAX_LEN ||
-        !settingsIsPrintableAscii(req.leaderName, 0x20)) {
-      request->send(400, "text/plain", F("Invalid leaderName"));
-      return;
-    }
-    // Source-IP binding (#313): a join mints a membership pointing display
-    // and firmware traffic at leaderHost, so the caller must actually BE
-    // leaderHost — the real leader dials from WiFi.localIP(), which is the
-    // exact value it puts in this field. A CSRF'd browser or any other LAN
-    // host cannot satisfy this, so it cannot hijack the row/membership.
-    if (req.leaderHost != request->client()->remoteIP().toString()) {
-      request->send(403, "text/plain",
-                    F("leaderHost must match the caller's address"));
-      return;
-    }
-    // #295 sticky leadership: a board that LEADS a wall never becomes a
-    // row of someone else's — after a promote, the returning old leader's
-    // joins collect this marker (from us and from every claimed member)
-    // and it demotes instead of building a second cluster.
-    if (clusterLeaderEnabled()) {
-      String out = "{\"error\":\"other-leader\",\"leaderHost\":";
-      appendJsonString(out, WiFi.localIP().toString());
-      out += ",\"leaderName\":";
-      appendJsonString(out, effectiveName);
-      out += '}';
-      request->send(409, "application/json", out);
-      return;
-    }
-    // While our current leader is demonstrably alive, a DIFFERENT
-    // leader's join is refused with its identity — the rejected board
-    // demotes itself on this marker.
-    if (clusterFollowerJoinWouldConflict(req.leaderHost)) {
-      ClusterFollowerView cv = clusterFollowerViewGet();
-      String out = "{\"error\":\"other-leader\",\"leaderHost\":";
-      appendJsonString(out, cv.leaderHost);
-      out += ",\"leaderName\":";
-      appendJsonString(out, cv.leaderName);
-      out += '}';
-      request->send(409, "application/json", out);
-      return;
-    }
-    clusterFollowerHandleJoin(req);
-    // Handshake reply: identity, firmware rev, width. Width is the boot
-    // probe's result today; #234 refines it — no protocol change. The
-    // #294 health keys ride along (minus width/rev, already present) so
-    // the leader's strip is live from the handshake, not the first ping.
-    DisplaySnapshot snap = displaySnapshotGet();
-    char mask[16];
-    clusterFaultMaskHex(snap.units, snap.displayWidth, mask, sizeof(mask));
-    WearAssessment wear;
-    assessWear(snap.units, snap.displayWidth, wear);
-    String out = "{\"name\":";
-    appendJsonString(out, effectiveName);
-    out += ",\"rev\":\"" GIT_REV "\",\"width\":";
-    out += (int)snap.displayWidth;
-    out += ",\"detected\":";
-    out += (int)snap.detectedUnitCount;
-    out += ",\"faulty\":";
-    out += (int)snap.faultyUnitCount;
-    out += ",\"faultMask\":\"";
-    out += mask;
-    out += "\",\"wear\":";
-    out += wear.flaggedCount > 0 ? "true" : "false";
-    out += ",\"protocol\":1}";
-    request->send(200, "application/json", out);
-  });
-
-  server.on("/cluster/render", HTTP_POST, [](AsyncWebServerRequest* request) {
-    // Source-IP binding (#313): only the joined leader drives this row's
-    // screen. This endpoint is never called by any browser UI (leader wire
-    // only), so a strict bind is safe. Standalone (leaderHost "") falls
-    // through to the handler's NotClustered 409.
-    ClusterFollowerView cv = clusterFollowerViewGet();
-    if (cv.leaderHost.length() > 0 &&
-        cv.leaderHost != request->client()->remoteIP().toString()) {
-      request->send(403, "text/plain", F("render must come from the leader"));
-      return;
-    }
-    if (!request->hasParam("epoch", true) || !request->hasParam("seq", true) ||
-        !request->hasParam("text", true)) {
-      request->send(400, "text/plain", F("Missing epoch/seq/text"));
-      return;
-    }
-    uint32_t epoch = (uint32_t)strtoul(
-        request->getParam("epoch", true)->value().c_str(), nullptr, 10);
-    uint32_t seq = (uint32_t)strtoul(
-        request->getParam("seq", true)->value().c_str(), nullptr, 10);
-    String text = request->getParam("text", true)->value();
-    int speed;
-    if (request->hasParam("speed", true)) {
-      speed = request->getParam("speed", true)->value().toInt();
-      if (speed < 1 || speed > 100) {
-        request->send(400, "text/plain", F("Speed must be 1..100"));
-        return;
-      }
-    } else {
-      WebStateLock lock;
-      speed = liveSettings->flapSpeed;
-    }
-    uint64_t commitAtMs =
-        request->hasParam("commitAtMs", true)
-            ? strtoull(request->getParam("commitAtMs", true)->value().c_str(),
-                       nullptr, 10)
-            : 0ULL;
-    // Wire-auth (#313 follow-on): once a key is negotiated, the render must
-    // carry a valid ts+mac over its exact content — a spoofed-IP host without
-    // the key cannot forge it, and the mac binds the text so it can't be
-    // swapped. Rebuild the canonical message from the wire params.
-    if (clusterFollowerHmacEnforced()) {
-      if (!request->hasParam("ts", true) || !request->hasParam("mac", true)) {
-        request->send(403, "text/plain", F("cluster signature required"));
-        return;
-      }
-      uint64_t ts = strtoull(request->getParam("ts", true)->value().c_str(),
-                             nullptr, 10);
-      String mac = request->getParam("mac", true)->value();
-      String msg = clusterHmacRenderMsg(ts, epoch, seq, text, speed, commitAtMs);
-      if (!clusterFollowerVerifySigned(msg, ts, mac)) {
-        request->send(403, "text/plain", F("bad cluster signature"));
-        return;
-      }
-    }
-    ClusterRenderVerdict v =
-        clusterFollowerHandleRender(epoch, seq, text, speed, commitAtMs);
-    if (v == ClusterRenderVerdict::NotClustered) {
-      request->send(409, "application/json",
-                    F("{\"error\":\"not clustered\"}"));
-      return;
-    }
-    String out = "{\"applied\":";
-    out += (v == ClusterRenderVerdict::Apply) ? "true" : "false";
-    out += ",\"seq\":";
-    out += String((unsigned long)seq);
-    out += '}';
-    request->send(200, "application/json", out);
-  });
-
-  server.on("/cluster/ping", HTTP_POST, [](AsyncWebServerRequest* request) {
-    // #294: the leader piggybacks the cluster digest (+ this member's
-    // table index) on the ping body, and the reply carries this row's
-    // unit health — both additive; either side may predate the other.
-    String digest = request->hasParam("digest", true)
-                        ? request->getParam("digest", true)->value()
-                        : String();
-    int youIndex = request->hasParam("you", true)
-                       ? (int)request->getParam("you", true)->value().toInt()
-                       : -1;
-    // Wire-auth (#313 follow-on): a keyed follower requires a valid ts+mac
-    // before the digest/you piggyback is trusted or contact is refreshed.
-    if (clusterFollowerHmacEnforced()) {
-      if (!request->hasParam("ts", true) || !request->hasParam("mac", true)) {
-        request->send(403, "text/plain", F("cluster signature required"));
-        return;
-      }
-      uint64_t ts = strtoull(request->getParam("ts", true)->value().c_str(),
-                             nullptr, 10);
-      String mac = request->getParam("mac", true)->value();
-      if (!clusterFollowerVerifySigned(clusterHmacPingMsg(ts, digest, youIndex),
-                                       ts, mac)) {
-        request->send(403, "text/plain", F("bad cluster signature"));
-        return;
-      }
-    }
-    if (!clusterFollowerHandlePing(digest, youIndex,
-                                   request->client()->remoteIP().toString())) {
-      request->send(409, "application/json",
-                    F("{\"error\":\"not clustered\"}"));
-      return;
-    }
-    ClusterFollowerView cv = clusterFollowerViewGet();
-    DisplaySnapshot snap = displaySnapshotGet();
-    WearAssessment wear;
-    assessWear(snap.units, snap.displayWidth, wear);
-    String out = "{\"state\":";
-    appendJsonString(out, clusterFollowerPhaseName(cv.phase));
-    out += ",\"epoch\":";
-    out += String((unsigned long)cv.epoch);
-    out += ",\"seq\":";
-    out += String((unsigned long)cv.lastSeq);
-    out += clusterPingHealthJson(snap.units, snap.displayWidth,
-                                 snap.detectedUnitCount, snap.faultyUnitCount,
-                                 wear.flaggedCount > 0, GIT_REV);
-    out += '}';
-    request->send(200, "application/json", out);
-  });
-
-  // #294 rung 2: the stored ping-piggybacked digest — any pane renders the
-  // whole wall from it. 404 = not clustered / nothing held yet (the page
-  // falls back to the standalone view).
-  server.on("/cluster/digest", HTTP_GET, [](AsyncWebServerRequest* request) {
-    uint32_t ageMs = 0;
-    String digest = clusterFollowerDigestGet(ageMs);
-    if (digest.length() == 0) {
-      request->send(404, "application/json", F("{\"error\":\"no digest\"}"));
-      return;
-    }
-    String out = "{\"ageMs\":";
-    out += String((unsigned long)ageMs);
-    out += ",\"digest\":";
-    out += digest;  // raw JSON from the leader — embedded, not escaped
-    out += '}';
-    request->send(200, "application/json", out);
-  });
-
-  // #295: one-click takeover from a follower that has written the leader
-  // off. The staged config swap runs in clusterTask; sticky-leadership
-  // join 409s resolve any promote race.
-  server.on("/cluster/promote", HTTP_POST,
-            [](AsyncWebServerRequest* request) {
-              ClusterPromoteVerdict v = clusterFollowerPromote();
-              String out = "{\"message\":";
-              appendJsonString(out, v.message);
-              out += '}';
-              request->send(v.httpStatus, "application/json", out);
-            });
-
-  server.on("/cluster/leave", HTTP_POST, [](AsyncWebServerRequest* request) {
-    // Combined guard (#313): /cluster/leave has TWO legitimate callers — the
-    // leader's reconfigure fan-out (server-to-server, no Origin, dials from
-    // leaderHost) and the local "Leave" button in the Cluster card (a LAN
-    // browser). So honor a leave from the leader's IP OR from a LAN-origin
-    // browser; a bare non-leader LAN host (no Origin, wrong IP) is refused,
-    // closing the any-host force-leave DoS without breaking the UI button.
-    // (The CSRF middleware has already 403'd any non-LAN browser origin.)
-    ClusterFollowerView cv = clusterFollowerViewGet();
-    bool fromLanBrowser = request->hasHeader("Origin") &&
-                          clusterCorsOriginAllowed(request->header("Origin"));
-    if (clusterFollowerHmacEnforced()) {
-      // Keyed (#313 follow-on): the leader arm becomes a valid SIGNATURE
-      // (beats a spoofed IP); the local Leave button rides the LAN-browser
-      // arm. A bare unsigned non-browser leave is refused.
-      bool signedOk = false;
-      if (request->hasParam("ts", true) && request->hasParam("mac", true)) {
-        uint64_t ts = strtoull(request->getParam("ts", true)->value().c_str(),
-                               nullptr, 10);
-        signedOk = clusterFollowerVerifySigned(
-            clusterHmacLeaveMsg(ts), ts,
-            request->getParam("mac", true)->value());
-      }
-      if (!signedOk && !fromLanBrowser) {
-        request->send(403, "text/plain",
-                      F("leave requires a valid signature or this display's "
-                        "own web UI"));
-        return;
-      }
-    } else {
-      // Pre-HMAC combined guard (#313): the leader's reconfigure fan-out
-      // (no Origin, dials from leaderHost) OR the local browser; a bare
-      // non-leader LAN host is refused (closes the any-host force-leave DoS).
-      bool fromLeader =
-          cv.leaderHost.length() > 0 &&
-          cv.leaderHost == request->client()->remoteIP().toString();
-      if (cv.leaderHost.length() > 0 && !fromLeader && !fromLanBrowser) {
-        request->send(403, "text/plain", F("leave must come from the leader "
-                                           "or this display's own web UI"));
-        return;
-      }
-    }
-    clusterFollowerHandleLeave();  // idempotent
-    request->send(200, "text/plain", F("ok"));
-  });
-
-  server.on("/cluster/health", HTTP_GET, [](AsyncWebServerRequest* request) {
-    ClusterFollowerView cv = clusterFollowerViewGet();
-    DisplaySnapshot snap = displaySnapshotGet();
-    String out = "{\"state\":";
-    appendJsonString(out, clusterFollowerPhaseName(cv.phase));
-    out += ",\"leaderName\":";
-    appendJsonString(out, cv.leaderName);
-    out += ",\"leaderHost\":";
-    appendJsonString(out, cv.leaderHost);
-    out += ",\"row\":";
-    out += cv.row;
-    out += ",\"epoch\":";
-    out += String((unsigned long)cv.epoch);
-    out += ",\"seq\":";
-    out += String((unsigned long)cv.lastSeq);
-    out += ",\"segment\":";
-    appendJsonString(out, cv.heldSegment);
-    out += ",\"rev\":\"" GIT_REV "\",\"width\":";
-    out += (int)snap.displayWidth;
-    out += ",\"detected\":";
-    out += (int)snap.detectedUnitCount;
-    out += ",\"faulty\":";
-    out += (int)snap.faultyUnitCount;
-    out += ",\"hmac\":";  // #313 follow-on: enforcing signed leader-wire requests
-    out += clusterFollowerHmacEnforced() ? "true" : "false";
-    out += '}';
-    request->send(200, "application/json", out);
-  });
-
-  // --- Cluster leader endpoints (#273) ----------------------------------------
-  // `members` uses the ClusterLeaderPolicy wire format
-  // (`host|row|col|width;…`, empty host = this master's own row; "" =
-  // disable). Validation runs here for the 400; the swap itself (leave
-  // fan-out, NVS persist, runtime reset) runs in clusterTask.
-  server.on("/cluster/config", HTTP_POST, [](AsyncWebServerRequest* request) {
-    if (!request->hasParam("members", true)) {
-      request->send(400, "text/plain", F("Missing members"));
-      return;
-    }
-    ClusterConfigVerdict v = clusterLeaderStageConfig(
-        request->getParam("members", true)->value());
-    request->send(v.httpStatus, "text/plain", v.message);
-  });
-
-  // Board discovery for the Cluster card (#274): POST arms the staged
-  // flag (re-POST while pending is a no-op — the flag is the re-entry
-  // guard); the blocking MDNS.queryService pass runs from netTask's
-  // drain, never here (/mqtt/discover contract).
-  server.on("/cluster/discover", HTTP_POST,
-            [](AsyncWebServerRequest* request) {
-    {
-      WebStateLock lock;
-      if (!clusterDiscoverPending) {
-        clusterDiscoverResultJson = "";
-        clusterDiscoverPending = true;
-      }
-    }
-    request->send(200, "text/plain", F("Board discovery started"));
-  });
-  server.on("/cluster/discover", HTTP_GET,
-            [](AsyncWebServerRequest* request) {
-    bool pending;
-    String json;
-    {
-      WebStateLock lock;
-      pending = clusterDiscoverPending;
-      json = clusterDiscoverResultJson;
-    }
-    if (pending) {
-      request->send(202, "text/plain", F("Discovery running"));
-      return;
-    }
-    if (json.length() == 0) {
-      request->send(404, "text/plain", F("No discovery has run yet"));
-      return;
-    }
-    request->send(200, "application/json", json);
-  });
-
-  server.on("/cluster/status", HTTP_GET, [](AsyncWebServerRequest* request) {
-    // One wire shape for the leader's status (#273 + #294 health keys),
-    // shared with the ping digest — serializer in ClusterDigest.h.
-    request->send(200, "application/json",
-                  clusterStatusJson(clusterLeaderStatusGet()));
-  });
-
-  // --- ESP-01 follower firmware: store + relay (#304 Part B) ------------------
-  // Upload a follower-<rev>.bin ONCE to the S3 (same-origin — no CORS); it's
-  // held on `storage` LittleFS and streamed on demand to esp01 rows by
-  // clusterTask, so the wall never browser-uploads to a follower (keeps #294's
-  // /firmware/* closed). Same async-context exception as the master OTA: the
-  // stream is accumulated in PSRAM here, MD5-verified, then handed to netTask.
-  // The follower-*.bin prefix guard mirrors ota-flash.sh #299 (an S3 image
-  // bricks the ESP-01).
-  server.on(
-      "/cluster/follower-firmware", HTTP_POST,
-      [](AsyncWebServerRequest* request) {
-        if (followerFwRejectStatus != 0) {
-          int s = followerFwRejectStatus;
-          String r = followerFwRejectReason;
-          followerFwRejectStatus = 0;
-          followerFwRejectReason = "";
-          request->send(s, "text/plain", r);
-          return;
-        }
-        if (!followerImageWriteEnd()) {
-          request->send(500, "text/plain",
-                        "Follower image store failed: " +
-                            followerImageWriteError());
-          return;
-        }
-        request->send(200, "text/plain",
-                      F("Follower image stored — use ‘Update firmware’ "
-                        "on an ESP-01 member to push it."));
-      },
-      [](AsyncWebServerRequest* request, String filename, size_t index,
-         uint8_t* data, size_t len, bool final) {
-        if (index == 0) {
-          followerFwRejectStatus = 0;
-          followerFwRejectReason = "";
-          // CSRF gate (#313), INLINE before followerImageWriteBegin — the
-          // middleware fires post-body, too late for an upload route. This
-          // upload is same-origin from the board's own UI, so its LAN Origin
-          // passes; a forged cross-site push is refused.
-          if (webUploadCsrfRejected(request)) {
-            followerFwRejectStatus = 403;
-            followerFwRejectReason =
-                "Cross-origin follower-image upload refused (CSRF guard)";
-            return;
-          }
-          String rev;
-          if (!followerImageUploadAccepts(filename, rev)) {
-            followerFwRejectStatus = 400;
-            followerFwRejectReason =
-                "expected a follower-<rev>.bin (an S3 image would brick the "
-                "ESP-01)";
-            return;
-          }
-          String md5 = request->hasParam("md5")
-                           ? request->getParam("md5")->value()
-                           : String();
-          if (md5.length() == 0) {
-            followerFwRejectStatus = 400;
-            followerFwRejectReason = "md5 query parameter is required";
-            return;
-          }
-          if (!normalizeOtaMd5(md5)) {
-            followerFwRejectStatus = 400;
-            followerFwRejectReason = "md5 must be exactly 32 hex characters";
-            return;
-          }
-          if (!followerImageWriteBegin(md5, rev)) {
-            followerFwRejectStatus = 409;
-            followerFwRejectReason =
-                "cannot start: " + followerImageWriteError();
-            return;
-          }
-        }
-        if (followerFwRejectStatus != 0) return;
-        if (len > 0 && !followerImageWriteChunk(data, len, index)) {
-          followerFwRejectStatus = 400;
-          followerFwRejectReason =
-              "upload failed: " + followerImageWriteError();
-          return;
-        }
-        // final: the completion handler calls followerImageWriteEnd().
-      });
-
-  // On-demand: stream the stored image to one esp01 member. Validation +
-  // eligibility live in clusterLeaderStageFollowerPush; the file stream runs
-  // on clusterTask (never here). Progress rides GET /cluster/status.
-  server.on("/cluster/member/update", HTTP_POST,
-            [](AsyncWebServerRequest* request) {
-    String host = request->hasParam("host", true)
-                      ? request->getParam("host", true)->value()
-                  : request->hasParam("host")
-                      ? request->getParam("host")->value()
-                      : String();
-    ClusterConfigVerdict v = clusterLeaderStageFollowerPush(host);
-    request->send(v.httpStatus, "text/plain", v.message);
-  });
+  // Per-module route registration (#338). Cross-module order is not
+  // semantic — the server matches per path+method, same-path method pairs
+  // stay within one module, and onNotFound/middleware are setters.
+  webContentRegister(server);
+  webSettingsRegister(server);
+  webSystemRegister(server);
+  webFirmwareRegister(server);
+  webMaintenanceRegister(server);
+  webClusterRegister(server);
 }
 
 void webEndpointsStart(AsyncWebServer& server) {
@@ -1982,60 +198,16 @@ void webEndpointsStart(AsyncWebServer& server) {
   SerialPrintln(F("Web server started"));
 }
 
-void webDisplayEventsTick() {
-  // 100 ms check cadence: imperceptible latency, and the snapshot copy
-  // stays off netTask's 10 ms hot loop.
-  static uint32_t nextCheckMs = 0;
-  static DisplayEventTracker tracker;
-  uint32_t nowMs = millis();
-  if ((int32_t)(nowMs - nextCheckMs) < 0) return;
-  nextCheckMs = nowMs + 100;
-
-  // count() before consuming the change: the library runs onConnect before
-  // inserting the client into its list, so a change landing in that window
-  // must stay unconsumed for the next tick. The resulting duplicate push of
-  // an already-shown text is deduped by the mirror's frame compare.
-  if (sseEvents.count() == 0) return;
-  DisplaySnapshot snap = displaySnapshotGet();
-
-  // Cheap pre-check before paying for the wall reconstruction (mutex +
-  // String churn, every 100 ms otherwise): only rebuild when the own text
-  // changed, the grid generation moved, or leading flipped. The rows-key
-  // due-check below stays authoritative for what actually gets pushed.
-  static uint32_t lastGridGen = 0;
-  static bool lastLeading = false;
-  bool leading = clusterLeaderEnabled();
-  uint32_t gridGen = leading ? clusterLeaderGridGeneration() : 0;
-  if (strncmp(tracker.lastText, snap.currentText, DISPLAY_CMD_TEXT_LEN) == 0 &&
-      leading == lastLeading && gridGen == lastGridGen) {
-    return;
-  }
-  lastLeading = leading;
-  lastGridGen = gridGen;
-
-  String rowsKey;
-  String payload = sseDisplayPayload(snap, &rowsKey);
-  if (!displayEventDue(tracker, snap.currentText, rowsKey)) return;
-  sseEvents.send(payload.c_str(), "display", millis());
-}
-
 void webEndpointsLoop(MasterSettings& settings, SettingsStore& store) {
   if (webStateMutex == nullptr) return;  // init hasn't run
 
-  // OTA stall watchdog (#313): a session that opened /firmware/master and
-  // went silent has frozen MQTT (mqttStopForOta) and holds the Update slot.
-  // If no chunk has landed for the timeout, thaw everything — the next
-  // upload's Update.begin() recovers the abandoned session (v1 #191). Mirror
-  // of the ESP-01 follower's webOtaUploadFrozen(); the stalled peer is parked
-  // in socket-read, not writing flash, so there is no concurrent Update use.
-  if (otaOwnerRequest != nullptr &&
-      millis() - otaLastChunkMs > OTA_STALL_TIMEOUT_MS) {
-    SerialPrintln(F("Master OTA upload stalled >30 s — aborting and resuming "
-                    "normal operation"));
-    if (Update.isRunning()) Update.abort();
-    otaOwnerRequest = nullptr;
-    mqttResumeAfterOta();
-  }
+  // Drain order is LOAD-BEARING and preserved verbatim from the pre-#338
+  // monolith: (1) OTA stall watchdog, (2) settings-post drain under the
+  // lock, (3) reboot hold fan-out, (4) rebootless TZ apply, (5) MQTT then
+  // (6) cluster mDNS discovery (LWIP locks — never nested inside
+  // webStateMutex), (7) flash-log + follower-image flushes (netTask is the
+  // sole flash writer), (8) the reboot itself last.
+  webFirmwareLoop();
 
   // The whole drain holds the mutex: applySettingsPost mutates the same
   // `settings` Strings GET /settings snapshots from the async task, and its
@@ -2199,84 +371,10 @@ void webEndpointsLoop(MasterSettings& settings, SettingsStore& store) {
   // lock domains from ever nesting (v1 #48 parity: TZ applies rebootless).
   if (timezoneChanged) clockServiceApplyTz(settings);
 
-  // MQTT broker discovery (#224): the blocking MDNS.queryService pass runs
-  // out here for the same lock-domain reason — mDNS takes LWIP locks. The
-  // ~1-2 s stall of netTask while it runs is the v1 loop() behavior.
-  bool discoverDue;
-  {
-    WebStateLock lock;
-    discoverDue = mqttDiscoverPending;
-  }
-  if (discoverDue) {
-    MdnsBrokerCandidate candidates[4];
-    size_t count = 0;
-    int n = MDNS.queryService("mqtt", "tcp");
-    for (int i = 0; i < n && count < 4; i++) {
-      MdnsBrokerCandidate& c = candidates[count++];
-      c.name = normalizeMdnsHostname(MDNS.hostname(i));
-      IPAddress a = MDNS.address(i);
-      c.ip = a == IPAddress() ? String() : a.toString();
-      c.advertisedPort = MDNS.port(i);
-      c.fromHomeAssistant = false;
-    }
-    if (count == 0) {
-      // No broker advertised itself — Home Assistant's zeroconf record
-      // locates the host running the Mosquitto add-on (v1 fallback).
-      n = MDNS.queryService("home-assistant", "tcp");
-      for (int i = 0; i < n && count < 4; i++) {
-        MdnsBrokerCandidate& c = candidates[count++];
-        c.name = normalizeMdnsHostname(MDNS.hostname(i));
-        IPAddress a = MDNS.address(i);
-        c.ip = a == IPAddress() ? String() : a.toString();
-        c.advertisedPort = MDNS.port(i);
-        c.fromHomeAssistant = true;
-      }
-    }
-    String json = buildDiscoverJson(candidates, count);
-    SerialPrintf("MQTT discover: %u candidate(s)\n", (unsigned)count);
-    WebStateLock lock;
-    mqttDiscoverResultJson = json;
-    mqttDiscoverPending = false;
-  }
-
-  // Cluster board discovery (#274): same lock-domain rationale as the MQTT
-  // pass above — mDNS takes LWIP locks, so the blocking query runs out here.
-  bool clusterDiscoverDue;
-  {
-    WebStateLock lock;
-    clusterDiscoverDue = clusterDiscoverPending;
-  }
-  if (clusterDiscoverDue) {
-    ClusterDiscoveredBoard boards[CLUSTER_DISCOVER_MAX_BOARDS];
-    size_t count = 0;
-    int n = MDNS.queryService("splitflap", "tcp");
-    for (int i = 0; i < n && count < CLUSTER_DISCOVER_MAX_BOARDS; i++) {
-      // TXT name (what the board calls itself) over the answer hostname —
-      // identical today, but the TXT survives mDNS conflict renaming.
-      String txtName = MDNS.txt(i, "name");
-      String name = txtName.length() > 0
-                        ? txtName
-                        : normalizeMdnsHostname(MDNS.hostname(i));
-      // Self and nameless answers must not consume result slots — with a
-      // full wall the leader's own advertisement would otherwise crowd out
-      // the last real board (the JSON builder re-filters as backstop).
-      if (name.length() == 0 || name.equalsIgnoreCase(effectiveName)) {
-        continue;
-      }
-      ClusterDiscoveredBoard& b = boards[count++];
-      b.name = name;
-      IPAddress a = MDNS.address(i);
-      b.ip = a == IPAddress() ? String() : a.toString();
-      b.rev = MDNS.txt(i, "rev");
-      b.width = clusterParseTxtWidth(MDNS.txt(i, "width"));
-      b.plat = MDNS.txt(i, "plat");  // "" = S3 master (#297)
-    }
-    String json = buildClusterDiscoverJson(boards, count, effectiveName);
-    SerialPrintf("Cluster discover: %u board(s)\n", (unsigned)count);
-    WebStateLock lock;
-    clusterDiscoverResultJson = json;
-    clusterDiscoverPending = false;
-  }
+  // mDNS discovery drains (#224 MQTT, #274 cluster): blocking queries take
+  // LWIP locks, so they run out here in netTask, outside webStateMutex.
+  webSettingsDiscoverLoop();
+  webClusterDiscoverLoop();
 
   // Flash-log drain (#206): netTask is the single flash writer.
   flashLogTick(rebootDue);  // force on reboot so the last lines land
@@ -2356,9 +454,3 @@ String webTimezoneSnapshot() {
   WebStateLock lock;
   return liveSettings->timezonePosix;
 }
-
-// Bundled unit firmware accessors (#205) — see WebEndpoints.h for why this
-// TU is the only WebAssets.h includer. On the S3, PROGMEM is flash-mapped
-// and directly readable, so the pointer works as a plain byte buffer.
-const uint8_t* webUnitFirmwareBin() { return UNIT_FIRMWARE_BIN; }
-size_t webUnitFirmwareBinLen() { return UNIT_FIRMWARE_BIN_LEN; }
